% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_evaluate_dmgs_equation3p3.R
\name{bayesian_dq_4terms_v1}
\alias{bayesian_dq_4terms_v1}
\title{Evaluate DMGS equation 3.3}
\usage{
bayesian_dq_4terms_v1(lddi, lddd, mu1, pidopi1, pidopi2, mu2, dim)
}
\arguments{
\item{lddi}{inverse of second derivative of observed log-likelihood}

\item{lddd}{third derivative of observed log-likelihood}

\item{mu1}{DMGS mu1 vector}

\item{pidopi1}{first part of the prior term}

\item{pidopi2}{second part of the prior term}

\item{mu2}{DMGS mu2 matrix}

\item{dim}{number of parameters}
}
\description{
Evaluate DMGS equation 3.3
}
