% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{bnorm_p12}
\alias{bnorm_p12}
\title{Bootstrap}
\usage{
bnorm_p12(x, t1, t2, n)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{n}{number of random samples required}
}
\description{
Bootstrap
}
