% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{calc_revert2ml}
\alias{calc_revert2ml}
\title{determine revert2ml or not}
\usage{
calc_revert2ml(v5h, v6h, t3)
}
\arguments{
\item{v5h}{fifth parameter}

\item{v6h}{sixth parameter}

\item{t3}{a vector of predictors for the shape}
}
\value{
Logical
}
\description{
determine revert2ml or not
}
