% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42b_cauchy_libs.R
\name{cauchy_f1f}
\alias{cauchy_f1f}
\title{DMGS equation 3.3, f1 term}
\usage{
cauchy_f1f(y, v1, d1, v2, fd2)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, f1 term
}
