% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/042c_cauchy_derivs.R
\name{cauchy_fdd}
\alias{cauchy_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
cauchy_fdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
