% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42c_cauchy_derivs.R
\name{cauchy_lddda}
\alias{cauchy_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
cauchy_lddda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
