% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/042b_cauchy_libs.R
\name{cauchy_lmn}
\alias{cauchy_lmn}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
cauchy_lmn(x, v1, d1, v2, fd2, mm, nn)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the second derivative of the normalized log-likelihood
}
