% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/042b_cauchy_libs.R
\name{cauchy_logf}
\alias{cauchy_logf}
\title{Logf for RUST}
\usage{
cauchy_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
