% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42c_cauchy_derivs.R
\name{cauchy_logfdd}
\alias{cauchy_logfdd}
\title{Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
cauchy_logfdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
