% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42b_cauchy_libs.R
\name{cauchy_loglik}
\alias{cauchy_loglik}
\title{log-likelihood function}
\usage{
cauchy_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\description{
log-likelihood function
}
