% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/64c_cauchy_p1_derivs.R
\name{cauchy_p1_fdd}
\alias{cauchy_p1_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
cauchy_p1_fdd(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
