% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064b_cauchy_p1_libs.R
\name{cauchy_p1_lmn}
\alias{cauchy_p1_lmn}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
cauchy_p1_lmn(x, t, v1, d1, v2, d2, v3, fd3, mm, nn)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the second derivative of the normalized log-likelihood
}
