% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/64b_cauchy_p1_libs.R
\name{cauchy_p1_logf}
\alias{cauchy_p1_logf}
\title{Logf for RUST}
\usage{
cauchy_p1_logf(params, x, t)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\description{
Logf for RUST
}
