% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064b_cauchy_p1_libs.R
\name{cauchy_p1_loglik}
\alias{cauchy_p1_loglik}
\title{Cauchy-with-p1  observed log-likelihood function}
\usage{
cauchy_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar value.
}
\description{
Cauchy-with-p1  observed log-likelihood function
}
