% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{crhpflat_dmgs_cpmethod}
\alias{crhpflat_dmgs_cpmethod}
\title{Generates a comment about the method}
\usage{
crhpflat_dmgs_cpmethod()
}
\value{
String
}
\description{
Generates a comment about the method
}
