% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_example_data.R
\docType{data}
\name{d056pareto_p1k2_example_data_v1_x}
\alias{d056pareto_p1k2_example_data_v1_x}
\title{This is data to be included in my package
(Now adding the word value{} to avoid triggering CRAN tests)}
\description{
This is data to be included in my package
(Now adding the word value{} to avoid triggering CRAN tests)
}
