% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_example_data.R
\docType{data}
\name{d071frechet_p2k1_example_data_v1_t}
\alias{d071frechet_p2k1_example_data_v1_t}
\title{This is data to be included in my package
(Now adding the word value{} to avoid triggering CRAN tests)}
\description{
This is data to be included in my package
(Now adding the word value{} to avoid triggering CRAN tests)
}
