% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42b_cauchy_libs.R
\name{dcauchysub}
\alias{dcauchysub}
\title{Densities from MLE and RHP}
\usage{
dcauchysub(x, y, d1 = 0.01, fd2 = 0.01, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Densities from MLE and RHP
}
