% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{deriv_copyfdd}
\alias{deriv_copyfdd}
\title{Extract the results from derivatives and put them into f2}
\usage{
deriv_copyfdd(temp1, nx, dim)
}
\arguments{
\item{temp1}{output from derivative calculations}

\item{nx}{number of x values}

\item{dim}{number of parameters}
}
\value{
3d array
}
\description{
Extract the results from derivatives and put them into f2
}
