% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{deriv_copyldd}
\alias{deriv_copyldd}
\title{Extract the results from derivatives and put them into ldd}
\usage{
deriv_copyldd(temp1, nx, dim)
}
\arguments{
\item{temp1}{output from derivative calculations}

\item{nx}{number of x values}

\item{dim}{number of parameters}
}
\value{
Matrix
}
\description{
Extract the results from derivatives and put them into ldd
}
