% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71b_frechet_p2k1_libs.R
\name{dfrechet_p2k1}
\alias{dfrechet_p2k1}
\title{Frechet_k1-with-p2 density function}
\usage{
dfrechet_p2k1(x, t0, ymn, slope, lambda, log = FALSE, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{lambda}{the lambda parameter of the distribution}

\item{log}{logical for the density evaluation}

\item{kloc}{the known location parameter}
}
\value{
Vector
}
\description{
Frechet_k1-with-p2 density function
}
