% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053b_gev_k3_libs.R
\name{dgev_k3sub}
\alias{dgev_k3sub}
\title{Densities from MLE and RHP}
\usage{
dgev_k3sub(x, y, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{kshape}{the known shape parameter}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
