% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074b_gev_p1k3_libs.R
\name{dgev_p1k3}
\alias{dgev_p1k3}
\title{GEV-with-known-shape-with-p1 density function}
\usage{
dgev_p1k3(x, t0, ymn, slope, sigma, log = FALSE, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}

\item{log}{logical for the density evaluation}

\item{kshape}{the known shape parameter}
}
\value{
Vector
}
\description{
GEV-with-known-shape-with-p1 density function
}
