% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074b_gev_p1k3_libs.R
\name{dgev_p1k3sub}
\alias{dgev_p1k3sub}
\title{Densities from MLE and RHP}
\usage{
dgev_p1k3sub(x, t, y, t0, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{kshape}{the known shape parameter}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
