% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1n_libs.R
\name{dgev_p1n}
\alias{dgev_p1n}
\title{GEVD-with-p1: Density function}
\usage{
dgev_p1n(x, t0, params, log = FALSE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}

\item{log}{logical for the density evaluation}
}
\value{
Vector
}
\description{
GEVD-with-p1: Density function
}
