% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1n_libs.R
\name{dgev_p1nsub}
\alias{dgev_p1nsub}
\title{Densities for 5 predictions}
\usage{
dgev_p1nsub(x, t, y, t0, ics, minxi, maxxi, extramodels = FALSE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ics}{initial conditions for the maximum likelihood search}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{extramodels}{logical that indicates whether to add three additional prediction models}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities for 5 predictions
}
