% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{dgev_p1sub}
\alias{dgev_p1sub}
\title{Densities for 5 predictions}
\usage{
dgev_p1sub(
  x,
  t,
  y,
  t0,
  ics,
  d1 = 0.01,
  d2 = 0.01,
  fd3 = 0.01,
  d4 = 0.01,
  minxi,
  maxxi,
  extramodels = FALSE,
  aderivs = TRUE
)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ics}{initial conditions for the maximum likelihood search}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{extramodels}{a logical that indicates whether to add three additional prediction models}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Densities for 5 predictions
}
