% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{dhalfnormsub}
\alias{dhalfnormsub}
\title{Densities from MLE and RHP}
\usage{
dhalfnormsub(x, y, fd1 = 0.01, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Densities from MLE and RHP
}
