% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061b_lnorm_p1_libs.R
\name{dlnorm_p1sub}
\alias{dlnorm_p1sub}
\title{Densities from MLE and RHP}
\usage{
dlnorm_p1sub(x, t, y, t0, debug = FALSE, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{debug}{debug flag}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
