% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040b_logis_libs.R
\name{dlogis2sub}
\alias{dlogis2sub}
\title{Densities from MLE and RHP}
\usage{
dlogis2sub(x, y)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
