% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_evaluate_dmgs_equation3p3.R
\name{dmgs}
\alias{dmgs}
\title{Evaluate DMGS equation 3.3}
\usage{
dmgs(lddi, lddd, mu1, pidopi, mu2, dim)
}
\arguments{
\item{lddi}{inverse of second derivative of observed log-likelihood}

\item{lddd}{third derivative of observed log-likelihood}

\item{mu1}{DMGS mu1 vector}

\item{pidopi}{derivative of log prior}

\item{mu2}{DMGS mu2 matrix}

\item{dim}{number of parameters}
}
\description{
Evaluate DMGS equation 3.3
}
