% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60b_norm_p1_libs.R
\name{dnorm_p1}
\alias{dnorm_p1}
\title{Normal-with-p1 density function}
\usage{
dnorm_p1(x, t0, ymn, slope, sigma, log = FALSE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}

\item{log}{logical for the density evaluation}
}
\value{
Vector
}
\description{
Normal-with-p1 density function
}
