% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011b_pareto_k2_libs.R
\name{dpareto_k2_sub}
\alias{dpareto_k2_sub}
\title{Densities from MLE and RHP}
\usage{
dpareto_k2_sub(x, y, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{kscale}{the known scale parameter}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
