% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56b_pareto_p1k2_libs.R
\name{dpareto_p1k2}
\alias{dpareto_p1k2}
\title{pareto_k1-with-p2 density function}
\usage{
dpareto_p1k2(x, t0, ymn, slope, kscale, log = FALSE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{kscale}{the known scale parameter}

\item{log}{a logical for the density evaluation}
}
\description{
pareto_k1-with-p2 density function
}
