% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{dpareto_p1k2sub}
\alias{dpareto_p1k2sub}
\title{Densities from MLE and RHP}
\usage{
dpareto_p1k2sub(x, t, y, t0, kscale, debug = FALSE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{kscale}{the known scale parameter}

\item{debug}{debug flag}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
