% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25b_unif_libs.R
\name{dunif_formula}
\alias{dunif_formula}
\title{Predictive PDFs}
\usage{
dunif_formula(x, y)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}
}
\value{
Two vectors
}
\description{
Predictive PDFs
}
