% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73b_weibull_p2_libs.R
\name{dweibull_p2sub}
\alias{dweibull_p2sub}
\title{Densities from MLE and RHP}
\usage{
dweibull_p2sub(x, t, y, t0, fd1, d2, d3, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
