% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10b_exp_libs.R
\name{exp_f1f}
\alias{exp_f1f}
\title{DMGS equation 2.1, f1 term}
\usage{
exp_f1f(y, v1, fd1)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 2.1, f1 term
}
