% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10c_exp_derivs.R
\name{exp_fd}
\alias{exp_fd}
\title{First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
exp_fd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\description{
First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
