% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10c_exp_derivs.R
\name{exp_ldda}
\alias{exp_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
exp_ldda(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\description{
The second derivative of the normalized log-likelihood
}
