% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10b_exp_libs.R
\name{exp_lddd}
\alias{exp_lddd}
\title{Third derivative tensor of the log-likelihood}
\usage{
exp_lddd(x, v1, fd1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the log-likelihood
}
