% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010c_exp_derivs.R, R/10c_exp_derivs.R
\name{exp_lddda}
\alias{exp_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
exp_lddda(x, v1)

exp_lddda(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
3d array

3d array
}
\description{
The third derivative of the normalized log-likelihood

The third derivative of the normalized log-likelihood
}
