% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10b_exp_libs.R
\name{exp_logf}
\alias{exp_logf}
\title{Logf for RUST}
\usage{
exp_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\description{
Logf for RUST
}
