% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{exp_p1_ldd}
\alias{exp_p1_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
exp_p1_ldd(x, t, v1, d1, v2, d2)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
Second derivative matrix of the normalized log-likelihood
}
