% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55c_exp_p1_derivs.R
\name{exp_p1_lddda}
\alias{exp_p1_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
exp_p1_lddda(x, t, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
