% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{exp_p1_logf}
\alias{exp_p1_logf}
\title{Logf for RUST}
\usage{
exp_p1_logf(params, x, t)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
