% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055c_exp_p1_derivs.R
\name{exp_p1_logfddd}
\alias{exp_p1_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
exp_p1_logfddd(x, t, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
3d array
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
