% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{exp_p1_loglik}
\alias{exp_p1_loglik}
\title{observed log-likelihood function}
\usage{
exp_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar value.
}
\description{
observed log-likelihood function
}
