% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{exp_p1_means}
\alias{exp_p1_means}
\title{exp distribution: RHP mean}
\usage{
exp_p1_means(means, t0, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 2)
}
\arguments{
\item{means}{a logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\description{
exp distribution: RHP mean
}
