% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{exp_p1_mu1f}
\alias{exp_p1_mu1f}
\title{DMGS equation 3.3, mu1 term}
\usage{
exp_p1_mu1f(alpha, t0, v1, d1, v2, d2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, mu1 term
}
