% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55c_exp_p1_derivs.R
\name{exp_p1_mu2fa}
\alias{exp_p1_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
exp_p1_mu2fa(alpha, t, v1, v2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Minus the second derivative of the cdf, at alpha
}
