% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55c_exp_p1_derivs.R
\name{exp_p1_p1fa}
\alias{exp_p1_p1fa}
\title{The first derivative of the cdf}
\usage{
exp_p1_p1fa(x, t, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Vector
}
\description{
The first derivative of the cdf
}
