% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010c_exp_derivs.R, R/10c_exp_derivs.R
\name{exp_p1fa}
\alias{exp_p1fa}
\title{The first derivative of the cdf}
\usage{
exp_p1fa(x, v1)

exp_p1fa(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Vector

Vector
}
\description{
The first derivative of the cdf

The first derivative of the cdf
}
