% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10c_exp_derivs.R
\name{exp_pd}
\alias{exp_pd}
\title{First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
exp_pd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Vector
}
\description{
First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
