% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{findnt}
\alias{findnt}
\title{Find the number of predictors in the predictor}
\usage{
findnt(t)
}
\arguments{
\item{t}{a vector or matrix of predictors}
}
\value{
Vector
}
\description{
Find the number of predictors in the predictor
}
