% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{fixgevrange}
\alias{fixgevrange}
\title{Deal with situations in which the user wants d or p outside the GEV range}
\usage{
fixgevrange(y, v1, v2, v3)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
Deal with situations in which the user wants d or p outside the GEV range
}
