% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{fixgpdrange}
\alias{fixgpdrange}
\title{Deal with situations in which the user wants d or p outside the GPD range}
\usage{
fixgpdrange(y, v1, v2, v3)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
Deal with situations in which the user wants d or p outside the GPD range
}
