% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051b_frechet_k1_libs.R
\name{frechet_k1_f2f}
\alias{frechet_k1_f2f}
\title{DMGS equation 3.3, f2 term}
\usage{
frechet_k1_f2f(y, v1, fd1, v2, fd2, kloc)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\value{
3d array
}
\description{
DMGS equation 3.3, f2 term
}
