% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051c_frechet_k1_derivs.R
\name{frechet_k1_lddda}
\alias{frechet_k1_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
frechet_k1_lddda(x, v1, v2, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
