% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51b_frechet_k1_libs.R
\name{frechet_k1_logf}
\alias{frechet_k1_logf}
\title{Logf for RUST}
\usage{
frechet_k1_logf(params, x, kloc)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kloc}{the known location parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
