% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51c_frechet_k1_derivs.R
\name{frechet_k1_p1fa}
\alias{frechet_k1_p1fa}
\title{The first derivative of the cdf}
\usage{
frechet_k1_p1fa(x, v1, v2, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\description{
The first derivative of the cdf
}
