% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51b_frechet_k1_libs.R
\name{frechet_loglik}
\alias{frechet_loglik}
\title{log-likelihood function}
\usage{
frechet_loglik(vv, x, kloc)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{kloc}{the known location parameter}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
