% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051b_frechet_k1_libs.R
\name{frechet_logscores}
\alias{frechet_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
frechet_logscores(logscores, x, kloc)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{kloc}{the known location parameter}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
