% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51b_frechet_k1_libs.R
\name{frechet_means}
\alias{frechet_means}
\title{MLE and RHP predictive means}
\usage{
frechet_means(means, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 2, kloc)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}

\item{kloc}{the known location parameter}
}
\value{
Two scalars
}
\description{
MLE and RHP predictive means
}
