% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071c_frechet_p2k1_derivs.R
\name{frechet_p2k1_f1fw}
\alias{frechet_p2k1_f1fw}
\title{The first derivative of the density for WAIC}
\usage{
frechet_p2k1_f1fw(x, t, v1, v2, v3, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kloc}{the known location parameter}
}
\value{
Vector
}
\description{
The first derivative of the density for WAIC
}
