% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71c_frechet_p2k1_derivs.R
\name{frechet_p2k1_f2fa}
\alias{frechet_p2k1_f2fa}
\title{The second derivative of the density}
\usage{
frechet_p2k1_f2fa(x, t, v1, v2, v3, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kloc}{the known location parameter}
}
\description{
The second derivative of the density
}
