% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071b_frechet_p2k1_libs.R
\name{frechet_p2k1_lddd}
\alias{frechet_p2k1_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
frechet_p2k1_lddd(x, t, v1, d1, v2, d2, v3, fd3, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood
}
