% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071b_frechet_p2k1_libs.R
\name{frechet_p2k1_logf}
\alias{frechet_p2k1_logf}
\title{Logf for RUST}
\usage{
frechet_p2k1_logf(params, x, t, kloc)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kloc}{the known location parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
