% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71b_frechet_p2k1_libs.R
\name{frechet_p2k1_loglik}
\alias{frechet_p2k1_loglik}
\title{observed log-likelihood function}
\usage{
frechet_p2k1_loglik(vv, x, t, kloc)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kloc}{the known location parameter}
}
\value{
Scalar value.
}
\description{
observed log-likelihood function
}
