% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071b_frechet_p2k1_libs.R
\name{frechet_p2k1_means}
\alias{frechet_p2k1_means}
\title{frechet_k1 distribution: RHP mean}
\usage{
frechet_p2k1_means(
  means,
  t0,
  ml_params,
  lddi,
  lddd,
  lambdad_rhp,
  nx,
  dim,
  kloc
)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}

\item{kloc}{the known location parameter}
}
\value{
Two scalars
}
\description{
frechet_k1 distribution: RHP mean
}
