% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71b_frechet_p2k1_libs.R
\name{frechet_p2k1_mu2f}
\alias{frechet_p2k1_mu2f}
\title{DMGS equation 3.3, mu2 term}
\usage{
frechet_p2k1_mu2f(alpha, t0, v1, d1, v2, d2, v3, fd3, kloc)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\description{
DMGS equation 3.3, mu2 term
}
