% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71c_frechet_p2k1_derivs.R
\name{frechet_p2k1_mu2fa}
\alias{frechet_p2k1_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
frechet_p2k1_mu2fa(alpha, t, v1, v2, v3, kloc)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kloc}{the known location parameter}
}
\value{
Matrix
}
\description{
Minus the second derivative of the cdf, at alpha
}
