% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071b_frechet_p2k1_libs.R
\name{frechet_p2k1_p1f}
\alias{frechet_p2k1_p1f}
\title{DMGS equation 2.1, p1 term}
\usage{
frechet_p2k1_p1f(y, t0, v1, d1, v2, d2, v3, fd3, kloc)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\value{
Matrix
}
\description{
DMGS equation 2.1, p1 term
}
