% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71c_frechet_p2k1_derivs.R
\name{frechet_p2k1_pdd}
\alias{frechet_p2k1_pdd}
\title{Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
frechet_p2k1_pdd(x, t, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\description{
Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
