% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71b_frechet_p2k1_libs.R
\name{frechet_p2k1_predictordata}
\alias{frechet_p2k1_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
frechet_p2k1_predictordata(predictordata, x, t, t0, params, kloc)
}
\arguments{
\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}

\item{kloc}{the known location parameter}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
