% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100b_gamma_libs.R
\name{gamma_gg}
\alias{gamma_gg}
\title{Second derivative matrix of the expected log-likelihood}
\usage{
gamma_gg(v1, fd1, v2, fd2)
}
\arguments{
\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Square scalar matrix
}
\description{
Second derivative matrix of the expected log-likelihood
}
