% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100b_gamma_libs.R
\name{gamma_gmn}
\alias{gamma_gmn}
\title{One component of the second derivative of the expected log-likelihood}
\usage{
gamma_gmn(alpha, v1, fd1, v2, fd2, mm, nn)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\description{
One component of the second derivative of the expected log-likelihood
}
