% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100c_gamma_derivs.R
\name{gamma_lddda}
\alias{gamma_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gamma_lddda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
