% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100b_gamma_libs.R
\name{gamma_logf}
\alias{gamma_logf}
\title{Logf for RUST}
\usage{
gamma_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\description{
Logf for RUST
}
