% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100b_gamma_libs.R
\name{gamma_loglik}
\alias{gamma_loglik}
\title{log-likelihood function}
\usage{
gamma_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
