% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100b_gamma_libs.R
\name{gamma_mu1f}
\alias{gamma_mu1f}
\title{DMGS equation 3.3, mu1 term}
\usage{
gamma_mu1f(alpha, v1, fd1, v2, fd2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, mu1 term
}
