% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_boot}
\alias{gev_boot}
\title{Bootstrap}
\usage{
gev_boot(x, n)
}
\arguments{
\item{x}{a vector of training data values}

\item{n}{number of random samples required}
}
\value{
A list containing a matrix of simulated parameter values
}
\description{
Bootstrap
}
