% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_checkmle}
\alias{gev_checkmle}
\title{Check MLE}
\usage{
gev_checkmle(ml_params, minxi = -1, maxxi = 1)
}
\arguments{
\item{ml_params}{parameters}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}
}
\value{
No return value (just a message to the screen).
}
\description{
Check MLE
}
