% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110c_gev_derivs.R
\name{gev_f1fa}
\alias{gev_f1fa}
\title{The first derivative of the density}
\usage{
gev_f1fa(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
The first derivative of the density
}
