% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110c_gev_derivs.R
\name{gev_fd}
\alias{gev_fd}
\title{First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gev_fd(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
