% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_ggd_mev}
\alias{gev_ggd_mev}
\title{Derivative of expected information matrix, based on MEV routine gev.infomat}
\usage{
gev_ggd_mev(v1, d1, v2, fd2, v3, d3)
}
\arguments{
\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}
}
\value{
Square scalar matrix
}
\description{
Derivative of expected information matrix, based on MEV routine gev.infomat
}
