% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_ppm_temporary.R
\name{gev_k12_ppm_minusloglik}
\alias{gev_k12_ppm_minusloglik}
\title{Temporary dummy for one of the ppm models}
\usage{
gev_k12_ppm_minusloglik(x)
}
\arguments{
\item{x}{a vector of training data values}
}
\description{
Temporary dummy for one of the ppm models
}
