% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53b_gev_k3_libs.R
\name{gev_k3_f1f}
\alias{gev_k3_f1f}
\title{DMGS equation 3.3, f1 term}
\usage{
gev_k3_f1f(y, v1, d1, v2, fd2, kshape)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kshape}{the known shape parameter}
}
\value{
Matrix
}
\description{
DMGS equation 3.3, f1 term
}
