% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53c_gev_k3_derivs.R
\name{gev_k3_f1fa}
\alias{gev_k3_f1fa}
\title{The first derivative of the density}
\usage{
gev_k3_f1fa(x, v1, v2, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kshape}{the known shape parameter}
}
\value{
Vector
}
\description{
The first derivative of the density
}
