% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53b_gev_k3_libs.R
\name{gev_k3_ldd}
\alias{gev_k3_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
gev_k3_ldd(x, v1, d1, v2, fd2, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kshape}{the known shape parameter}
}
\value{
Square scalar matrix
}
\description{
Second derivative matrix of the normalized log-likelihood
}
