% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053b_gev_k3_libs.R
\name{gev_k3_lddd}
\alias{gev_k3_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
gev_k3_lddd(x, v1, d1, v2, fd2, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kshape}{the known shape parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood
}
