% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53b_gev_k3_libs.R
\name{gev_k3_logf}
\alias{gev_k3_logf}
\title{Logf for RUST}
\usage{
gev_k3_logf(params, x, kshape)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kshape}{the known shape parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
