% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053b_gev_k3_libs.R
\name{gev_k3_loglik}
\alias{gev_k3_loglik}
\title{log-likelihood function}
\usage{
gev_k3_loglik(vv, x, kshape)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{kshape}{the known shape parameter}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
