% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53c_gev_k3_derivs.R
\name{gev_k3_mu2fa}
\alias{gev_k3_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
gev_k3_mu2fa(alpha, v1, v2, kshape)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kshape}{the known shape parameter}
}
\description{
Minus the second derivative of the cdf, at alpha
}
