% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53c_gev_k3_derivs.R
\name{gev_k3_pd}
\alias{gev_k3_pd}
\title{First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gev_k3_pd(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
