% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110c_gev_derivs.R
\name{gev_ldda}
\alias{gev_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
gev_ldda(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
The second derivative of the normalized log-likelihood
}
