% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_lddd}
\alias{gev_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
gev_lddd(x, v1, d1, v2, fd2, v3, d3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
Third derivative tensor of the normalized log-likelihood
}
