% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110c_gev_derivs.R
\name{gev_lddda}
\alias{gev_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gev_lddda(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
