% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_logf}
\alias{gev_logf}
\title{Logf for RUST}
\usage{
gev_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\description{
Logf for RUST
}
