% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_means}
\alias{gev_means}
\title{Analytical Expressions for Predictive Means
RHP mean based on the expectation of DMGS equation 2.1}
\usage{
gev_means(
  means,
  ml_params,
  lddi,
  lddd,
  lambdad_rh_flat,
  lambdad_custom,
  nx,
  dim = 3
)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rh_flat}{derivative of the log CRHP-FLAT prior}

\item{lambdad_custom}{custom value of the derivative of the log prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
Analytical Expressions for Predictive Means
RHP mean based on the expectation of DMGS equation 2.1
}
