% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110c_gev_derivs.R
\name{gev_mu1fa}
\alias{gev_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gev_mu1fa(alpha, v1, v2, v3)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
Minus the first derivative of the cdf, at alpha
}
