% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_mu2f}
\alias{gev_mu2f}
\title{DMGS equation 3.3, mu2 term}
\usage{
gev_mu2f(alpha, v1, d1, v2, fd2, v3, d3)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, mu2 term
}
