% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{gev_p123_checkmle}
\alias{gev_p123_checkmle}
\title{Check MLE}
\usage{
gev_p123_checkmle(ml_params, minxi, maxxi, t1, t2, t3)
}
\arguments{
\item{ml_params}{parameters}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}
}
\value{
No return value (just a message to the screen).
}
\description{
Check MLE
}
