% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{gev_p123_f1f}
\alias{gev_p123_f1f}
\title{DMGS equation 2.1, f1 term, fixed shape parameter
DMGS equation 2.1, f1 term}
\usage{
gev_p123_f1f(y, t01, t02, t03, v1, d1, v2, d2, v3, d3, v4, d4, v5, d5, v6, d6)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{t03}{a single value of the predictor (specify either \code{t03} or \code{n03} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{v5}{fifth parameter}

\item{d5}{the delta used in the numerical derivatives with respect to the parameter}

\item{v6}{sixth parameter}

\item{d6}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 2.1, f1 term, fixed shape parameter
DMGS equation 2.1, f1 term
}
