% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/155c_gev_p123_derivs.R
\name{gev_p123_f2fa}
\alias{gev_p123_f2fa}
\title{The second derivative of the density for DMGS}
\usage{
gev_p123_f2fa(x, t01, t02, t03, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{t03}{a single value of the predictor (specify either \code{t03} or \code{n03} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density for DMGS
}
