% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{gev_p123_ldd}
\alias{gev_p123_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
gev_p123_ldd(x, t1, t2, t3, v1, d1, v2, d2, v3, d3, v4, d4, v5, d5, v6, d6)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{v5}{fifth parameter}

\item{d5}{the delta used in the numerical derivatives with respect to the parameter}

\item{v6}{sixth parameter}

\item{d6}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
Second derivative matrix of the normalized log-likelihood
}
