% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/155c_gev_p123_derivs.R
\name{gev_p123_ldda}
\alias{gev_p123_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
gev_p123_ldda(x, t1, t2, t3, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the normalized log-likelihood
}
