% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{gev_p123_logf}
\alias{gev_p123_logf}
\title{Logf for RUST}
\usage{
gev_p123_logf(params, x, t1, t2, t3)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
