% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152c_gev_p123_derivs.R
\name{gev_p123_logfddd}
\alias{gev_p123_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gev_p123_logfddd(x, t1, t2, t3, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
3d array
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
