% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/155b_gev_p123_libs.R
\name{gev_p123_loglik}
\alias{gev_p123_loglik}
\title{observed log-likelihood function}
\usage{
gev_p123_loglik(vv, x, t1, t2, t3)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}
}
\value{
Scalar
}
\description{
observed log-likelihood function
}
