% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/155b_gev_p123_libs.R
\name{gev_p123_predictordata}
\alias{gev_p123_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
gev_p123_predictordata(x, t1, t2, t3, t01, t02, t03, params)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{t03}{a single value of the predictor (specify either \code{t03} or \code{n03} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
