% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151c_gev_p12_derivs.R
\name{gev_p12_lddda}
\alias{gev_p12_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gev_p12_lddda(x, t1, t2, v1, v2, v3, v4, v5)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
