% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/154b_gev_p12_libs.R
\name{gev_p12_logf}
\alias{gev_p12_logf}
\title{Logf for RUST}
\usage{
gev_p12_logf(params, x, t1, t2)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
