% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151c_gev_p12_derivs.R
\name{gev_p12_mu1fa}
\alias{gev_p12_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gev_p12_mu1fa(alpha, t01, t02, v1, v2, v3, v4, v5)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}
}
\value{
Vector
}
\description{
Minus the first derivative of the cdf, at alpha
}
