% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151b_gev_p12_libs.R
\name{gev_p12_predictordata}
\alias{gev_p12_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
gev_p12_predictordata(predictordata, x, t1, t2, t01, t02, params)
}
\arguments{
\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
