% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151b_gev_p12_libs.R
\name{gev_p12_setics}
\alias{gev_p12_setics}
\title{Set initial conditions}
\usage{
gev_p12_setics(x, t1, t2, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{ics}{initial conditions for the maximum likelihood search}
}
\value{
Vector
}
\description{
Set initial conditions
}
