% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151b_gev_p12_libs.R
\name{gev_p12_waic}
\alias{gev_p12_waic}
\title{Waic}
\usage{
gev_p12_waic(
  waicscores,
  x,
  t1,
  t2,
  v1hat,
  d1,
  v2hat,
  d2,
  v3hat,
  d3,
  v4hat,
  d4,
  v5hat,
  d5,
  lddi,
  lddd,
  lambdad,
  aderivs
)
}
\arguments{
\item{waicscores}{a logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1hat}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2hat}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3hat}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{v4hat}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{v5hat}{fifth parameter}

\item{d5}{the delta used in the numerical derivatives with respect to the parameter}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad}{derivative of the log prior}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Waic
}
