% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074c_gev_p12k3_derivs.R
\name{gev_p12k3_mu2fa}
\alias{gev_p12k3_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
gev_p12k3_mu2fa(alpha, t01, t02, v1, v2, v3, v4, kshape)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{kshape}{the known shape parameter}
}
\value{
Matrix
}
\description{
Minus the second derivative of the cdf, at alpha
}
