% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{gev_p1_checkmle}
\alias{gev_p1_checkmle}
\title{Check MLE}
\usage{
gev_p1_checkmle(ml_params, minxi, maxxi)
}
\arguments{
\item{ml_params}{parameters}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}
}
\value{
No return value (just a message to the screen).
}
\description{
Check MLE
}
