% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{gev_p1_ggd}
\alias{gev_p1_ggd}
\title{Derivative of information matrix, based on ldd}
\usage{
gev_p1_ggd(x, t, v1, d1, v2, d2, v3, fd3, v4, d4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}
}
\value{
Square scalar matrix
}
\description{
Derivative of information matrix, based on ldd
}
