% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1_derivs.R
\name{gev_p1_mu1fa}
\alias{gev_p1_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gev_p1_mu1fa(alpha, t, v1, v2, v3, v4)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\description{
Minus the first derivative of the cdf, at alpha
}
