% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{gev_p1_setics}
\alias{gev_p1_setics}
\title{Set initial conditions}
\usage{
gev_p1_setics(x, t, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{ics}{initial conditions for the maximum likelihood search}
}
\description{
Set initial conditions
}
