% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153c_gev_p1a_derivs.R
\name{gev_p1a_f2fa}
\alias{gev_p1a_f2fa}
\title{The second derivative of the density for DMGS}
\usage{
gev_p1a_f2fa(x, t0, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density for DMGS
}
