% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1a_derivs.R
\name{gev_p1a_f2fw}
\alias{gev_p1a_f2fw}
\title{The second derivative of the density for WAIC}
\usage{
gev_p1a_f2fw(x, t, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density for WAIC
}
