% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1a_derivs.R
\name{gev_p1a_pdd}
\alias{gev_p1a_pdd}
\title{Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gev_p1a_pdd(x, t, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\value{
Matrix
}
\description{
Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
