% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1b_derivs.R
\name{gev_p1b_mu1fa}
\alias{gev_p1b_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gev_p1b_mu1fa(alpha, t0a, t0b, v1, v2, v3, v4, v5)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0a}{a single value of the predictor, for the first column of the predictor    (specify either \code{t0a} or \code{n0a} but not both)}

\item{t0b}{a single value of the predictor, for the second column of the predictor (specify either \code{t0b} or \code{n0b} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}
}
\value{
Vector
}
\description{
Minus the first derivative of the cdf, at alpha
}
