% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1c_derivs.R
\name{gev_p1c_f2fa}
\alias{gev_p1c_f2fa}
\title{The second derivative of the density for DMGS}
\usage{
gev_p1c_f2fa(x, t0a, t0b, t0c, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0a}{a single value of the predictor, for the first column of the predictor    (specify either \code{t0a} or \code{n0a} but not both)}

\item{t0b}{a single value of the predictor, for the second column of the predictor (specify either \code{t0b} or \code{n0b} but not both)}

\item{t0c}{a single value of the predictor, for the third column of the predictor    (specify either \code{t0c} or \code{n0c} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density for DMGS
}
