% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153c_gev_p1c_derivs.R
\name{gev_p1c_fdd}
\alias{gev_p1c_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gev_p1c_fdd(x, ta, tb, tc, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{ta}{a vector of predictors for the mean (first column)}

\item{tb}{a vector of predictors for the mean (second column)}

\item{tc}{a vector of predictors for the mean (third column)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
Matrix
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
