% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150c_gev_p1c_derivs.R
\name{gev_p1c_ldda}
\alias{gev_p1c_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
gev_p1c_ldda(x, ta, tb, tc, v1, v2, v3, v4, v5, v6)
}
\arguments{
\item{x}{a vector of training data values}

\item{ta}{a vector of predictors for the mean (first column)}

\item{tb}{a vector of predictors for the mean (second column)}

\item{tc}{a vector of predictors for the mean (third column)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}

\item{v5}{fifth parameter}

\item{v6}{sixth parameter}
}
\value{
Matrix
}
\description{
The second derivative of the normalized log-likelihood
}
