% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R, R/74b_gev_p1k3_libs.R
\name{gev_p1k3_lddd}
\alias{gev_p1k3_lddd}
\title{Third derivative tensor of the normalized log-likelihood, with fixed shape parameter}
\usage{
gev_p1k3_lddd(x, t, v1, d1, v2, d2, v3, fd3, kshape)

gev_p1k3_lddd(x, t, v1, d1, v2, d2, v3, fd3, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kshape}{the known shape parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood, with fixed shape parameter

Third derivative tensor of the normalized log-likelihood
}
