% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/74c_gev_p1k3_derivs.R
\name{gev_p1k3_lddda}
\alias{gev_p1k3_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gev_p1k3_lddda(x, t, v1, v2, v3, kshape)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kshape}{the known shape parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
