% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/74b_gev_p1k3_libs.R
\name{gev_p1k3_lmn}
\alias{gev_p1k3_lmn}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
gev_p1k3_lmn(x, t, v1, d1, v2, d2, v3, fd3, kshape, mm, nn)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kshape}{the known shape parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\description{
One component of the second derivative of the normalized log-likelihood
}
