% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074b_gev_p1k3_libs.R
\name{gev_p1k3_logf}
\alias{gev_p1k3_logf}
\title{Logf for RUST}
\usage{
gev_p1k3_logf(params, x, t, kshape)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kshape}{the known shape parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
