% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074b_gev_p1k3_libs.R
\name{gev_p1k3_loglik}
\alias{gev_p1k3_loglik}
\title{GEV-with-known-shape-with-p1  observed log-likelihood function}
\usage{
gev_p1k3_loglik(vv, x, t, kshape)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kshape}{the known shape parameter}
}
\value{
Scalar
}
\description{
GEV-with-known-shape-with-p1  observed log-likelihood function
}
