% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/74b_gev_p1k3_libs.R
\name{gev_p1k3_means}
\alias{gev_p1k3_means}
\title{Analytical expressions for Predictive Means
RHP mean based on the expectation of DMGS equation 2.1}
\usage{
gev_p1k3_means(means, t0, ml_params, kshape, nx)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ml_params}{parameters}

\item{kshape}{the known shape parameter}

\item{nx}{length of training data}
}
\value{
Two scalars
}
\description{
Analytical expressions for Predictive Means
RHP mean based on the expectation of DMGS equation 2.1
}
