% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1n_libs.R
\name{gev_p1n_logf}
\alias{gev_p1n_logf}
\title{Logf for RUST}
\usage{
gev_p1n_logf(params, x, t)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
