% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153b_gev_p1n_libs.R
\name{gev_p1n_loglik}
\alias{gev_p1n_loglik}
\title{observed log-likelihood function}
\usage{
gev_p1n_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar
}
\description{
observed log-likelihood function
}
