% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153b_gev_p1n_libs.R
\name{gev_p1n_n1_exampledata}
\alias{gev_p1n_n1_exampledata}
\title{GEV_p1n n=1 example data}
\usage{
gev_p1n_n1_exampledata(iseed)
}
\arguments{
\item{iseed}{The random seed}
}
\value{
A list containing data to run an example
}
\description{
GEV_p1n n=1 example data
}
