% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153b_gev_p1n_libs.R
\name{gev_p1n_predictordata}
\alias{gev_p1n_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
gev_p1n_predictordata(predictordata, x, t, t0, params)
}
\arguments{
\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
