% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1n_libs.R
\name{gev_p1n_setics}
\alias{gev_p1n_setics}
\title{Set initial conditions}
\usage{
gev_p1n_setics(x, t)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Vector
}
\description{
Set initial conditions
}
