% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153b_gev_p1n_libs.R
\name{gev_p1n_waic}
\alias{gev_p1n_waic}
\title{Waic}
\usage{
gev_p1n_waic(waicscores, x, t0, vhat, lddi, lddd, lambdad)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{vhat}{vector of all parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad}{derivative of the log prior}
}
\value{
Two numeric values.
}
\description{
Waic
}
