% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R, R/160b_gev_extramodels_libs.R
\name{gev_pwm_params}
\alias{gev_pwm_params}
\title{PWM parameter estimation}
\usage{
gev_pwm_params(x)

gev_pwm_params(x)
}
\arguments{
\item{x}{a vector of training data values}
}
\value{
Vector

Vector
}
\description{
PWM parameter estimation

PWM parameter estimation
}
