% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R
\name{gev_setics}
\alias{gev_setics}
\title{Set initial conditions}
\usage{
gev_setics(x, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{ics}{initial conditions for the maximum likelihood search}
}
\description{
Set initial conditions
}
