% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32c_gnorm_k3_derivs.R
\name{gnorm_k3_f1fa}
\alias{gnorm_k3_f1fa}
\title{The first derivative of the density}
\usage{
gnorm_k3_f1fa(x, v1, v2, kbeta)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kbeta}{the known beta parameter}
}
\description{
The first derivative of the density
}
