% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32b_gnorm_k3_libs.R
\name{gnorm_k3_f2f}
\alias{gnorm_k3_f2f}
\title{DMGS equation 3.3, f2 term}
\usage{
gnorm_k3_f2f(y, v1, d1, v2, fd2, kbeta)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kbeta}{the known beta parameter}
}
\value{
3d array
}
\description{
DMGS equation 3.3, f2 term
}
