% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32c_gnorm_k3_derivs.R
\name{gnorm_k3_lddda}
\alias{gnorm_k3_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gnorm_k3_lddda(x, v1, v2, kbeta)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kbeta}{the known beta parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
