% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32b_gnorm_k3_libs.R
\name{gnorm_k3_logf}
\alias{gnorm_k3_logf}
\title{Logf for RUST}
\usage{
gnorm_k3_logf(params, x, kbeta)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kbeta}{the known beta parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
