% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/032b_gnorm_k3_libs.R
\name{gnorm_k3_loglik}
\alias{gnorm_k3_loglik}
\title{log-likelihood function}
\usage{
gnorm_k3_loglik(vv, x, kbeta)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{kbeta}{the known beta parameter}
}
\value{
Scalar
}
\description{
log-likelihood function
}
