% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32b_gnorm_k3_libs.R
\name{gnorm_k3_logscores}
\alias{gnorm_k3_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
gnorm_k3_logscores(logscores, x, d1 = 0.01, fd2 = 0.01, kbeta, aderivs)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kbeta}{the known beta parameter}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
