% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32b_gnorm_k3_libs.R
\name{gnorm_k3_mu2f}
\alias{gnorm_k3_mu2f}
\title{DMGS equation 3.3, mu2 term}
\usage{
gnorm_k3_mu2f(alpha, v1, d1, v2, fd2, kbeta)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kbeta}{the known beta parameter}
}
\value{
3d array
}
\description{
DMGS equation 3.3, mu2 term
}
