% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k13_l111}
\alias{gpd_k13_l111}
\title{One component of the third derivative of the normalized log-likelihood}
\usage{
gpd_k13_l111(x, v1, fd1, v2, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\value{
Scalar value
}
\description{
One component of the third derivative of the normalized log-likelihood
}
