% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120c_gpd_k13_derivs.R
\name{gpd_k13_ldda}
\alias{gpd_k13_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
gpd_k13_ldda(x, v1, v2, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\description{
The second derivative of the normalized log-likelihood
}
