% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k13_mu2f}
\alias{gpd_k13_mu2f}
\title{DMGS equation 3.3, mu2 term}
\usage{
gpd_k13_mu2f(alpha, v1, fd1, v2, kloc)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\value{
3d array
}
\description{
DMGS equation 3.3, mu2 term
}
