% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120c_gpd_k13_derivs.R
\name{gpd_k13_pd}
\alias{gpd_k13_pd}
\title{First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
gpd_k13_pd(x, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
