% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_checkmle}
\alias{gpd_k1_checkmle}
\title{Check MLE}
\usage{
gpd_k1_checkmle(ml_params, kloc, minxi, maxxi)
}
\arguments{
\item{ml_params}{parameters}

\item{kloc}{the known location parameter}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}
}
\value{
No return value (just a message to the screen).
}
\description{
Check MLE
}
