% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120c_gpd_k1_derivs.R
\name{gpd_k1_f1fa}
\alias{gpd_k1_f1fa}
\title{The first derivative of the density}
\usage{
gpd_k1_f1fa(x, v1, v2, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\value{
Vector
}
\description{
The first derivative of the density
}
