% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_ggd_mev}
\alias{gpd_k1_ggd_mev}
\title{Derivative of expected information matrix, based on MEV routine gpd.infomat}
\usage{
gpd_k1_ggd_mev(v1, fd1, v2, d2, kloc)
}
\arguments{
\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\description{
Derivative of expected information matrix, based on MEV routine gpd.infomat
}
