% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_lmn}
\alias{gpd_k1_lmn}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
gpd_k1_lmn(x, v1, fd1, v2, d2, kloc, mm, nn)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\description{
One component of the second derivative of the normalized log-likelihood
}
