% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_logf}
\alias{gpd_k1_logf}
\title{Logf for RUST}
\usage{
gpd_k1_logf(params, x, kloc)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kloc}{the known location parameter}
}
\description{
Logf for RUST
}
