% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_loglik}
\alias{gpd_k1_loglik}
\title{log-likelihood function}
\usage{
gpd_k1_loglik(vv, x, kloc)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{kloc}{the known location parameter}
}
\description{
log-likelihood function
}
