% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_mu1f}
\alias{gpd_k1_mu1f}
\title{DMGS equation 3.3, mu1 term}
\usage{
gpd_k1_mu1f(alpha, v1, fd1, v2, d2, kloc)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kloc}{the known location parameter}
}
\value{
Matrix
}
\description{
DMGS equation 3.3, mu1 term
}
