% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120c_gpd_k1_derivs.R
\name{gpd_k1_mu1fa}
\alias{gpd_k1_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gpd_k1_mu1fa(alpha, v1, v2, kloc)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kloc}{the known location parameter}
}
\description{
Minus the first derivative of the cdf, at alpha
}
