% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_setics}
\alias{gpd_k1_setics}
\title{Set initial conditions}
\usage{
gpd_k1_setics(x, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{ics}{initial conditions for the maximum likelihood search}
}
\description{
Set initial conditions
}
