% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{gpd_k1_waic}
\alias{gpd_k1_waic}
\title{Waic}
\usage{
gpd_k1_waic(waicscores, x, v1hat, v2hat, kloc, lddi, lddd, lambdad)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}

\item{kloc}{the known location parameter}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad}{derivative of the log prior}
}
\value{
Two numeric values.
}
\description{
Waic
}
