% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050b_gumbel_libs.R
\name{gumbel_ldd}
\alias{gumbel_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
gumbel_ldd(x, v1, d1, v2, fd2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Square scalar matrix
}
\description{
Second derivative matrix of the normalized log-likelihood
}
