% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050c_gumbel_derivs.R
\name{gumbel_lddda}
\alias{gumbel_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
gumbel_lddda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
