% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50b_gumbel_libs.R
\name{gumbel_logf}
\alias{gumbel_logf}
\title{Logf for RUST}
\usage{
gumbel_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
