% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50b_gumbel_libs.R
\name{gumbel_loglik}
\alias{gumbel_loglik}
\title{log-likelihood function}
\usage{
gumbel_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
