% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050b_gumbel_libs.R
\name{gumbel_means}
\alias{gumbel_means}
\title{MLE and RHP predictive means}
\usage{
gumbel_means(means, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 2)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
MLE and RHP predictive means
}
