% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50c_gumbel_derivs.R
\name{gumbel_mu2fa}
\alias{gumbel_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
gumbel_mu2fa(alpha, v1, v2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Minus the second derivative of the cdf, at alpha
}
