% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070b_gumbel_p1_libs.R
\name{gumbel_p1_loglik}
\alias{gumbel_p1_loglik}
\title{observed log-likelihood function}
\usage{
gumbel_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar
}
\description{
observed log-likelihood function
}
