% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070b_gumbel_p1_libs.R
\name{gumbel_p1_logscores}
\alias{gumbel_p1_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
gumbel_p1_logscores(logscores, x, t)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
