% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070c_gumbel_p1_derivs.R
\name{gumbel_p1_mu1fa}
\alias{gumbel_p1_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
gumbel_p1_mu1fa(alpha, t0, v1, v2, v3)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
Minus the first derivative of the cdf, at alpha
}
