% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50c_gumbel_derivs.R
\name{gumbel_p2fa}
\alias{gumbel_p2fa}
\title{The second derivative of the cdf}
\usage{
gumbel_p2fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The second derivative of the cdf
}
