% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_f1fa}
\alias{halfnorm_f1fa}
\title{The first derivative of the density}
\usage{
halfnorm_f1fa(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Vector
}
\description{
The first derivative of the density
}
