% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_f2fa}
\alias{halfnorm_f2fa}
\title{The second derivative of the density}
\usage{
halfnorm_f2fa(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density
}
