% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_fd}
\alias{halfnorm_fd}
\title{First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
halfnorm_fd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Vector
}
\description{
First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
