% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_fdd}
\alias{halfnorm_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
halfnorm_fdd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Matrix
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
