% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_gg}
\alias{halfnorm_gg}
\title{Expected information matrix}
\usage{
halfnorm_gg(v1, fd1)
}
\arguments{
\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
Expected information matrix
}
