% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020b_halfnorm_libs.R
\name{halfnorm_gg11}
\alias{halfnorm_gg11}
\title{Second derivative of the expected log-likelihood}
\usage{
halfnorm_gg11(alpha, v1, fd1)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Scalar value
}
\description{
Second derivative of the expected log-likelihood
}
