% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_l111}
\alias{halfnorm_l111}
\title{Third derivative of the normalized log-likelihood}
\usage{
halfnorm_l111(x, v1, fd1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
Third derivative of the normalized log-likelihood
}
