% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020b_halfnorm_libs.R
\name{halfnorm_ldd}
\alias{halfnorm_ldd}
\title{The second derivative of the normalized log-likelihood}
\usage{
halfnorm_ldd(x, v1, fd1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Square scalar matrix
}
\description{
The second derivative of the normalized log-likelihood
}
