% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020c_halfnorm_derivs.R, R/20c_halfnorm_derivs.R
\name{halfnorm_ldda}
\alias{halfnorm_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
halfnorm_ldda(x, v1)

halfnorm_ldda(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Matrix

Matrix
}
\description{
The second derivative of the normalized log-likelihood

The second derivative of the normalized log-likelihood
}
