% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_lddd}
\alias{halfnorm_lddd}
\title{Third derivative tensor of the log-likelihood}
\usage{
halfnorm_lddd(x, v1, fd1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the log-likelihood
}
