% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_lddda}
\alias{halfnorm_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
halfnorm_lddda(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
