% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_logf}
\alias{halfnorm_logf}
\title{Logf for RUST}
\usage{
halfnorm_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
