% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020c_halfnorm_derivs.R, R/20c_halfnorm_derivs.R
\name{halfnorm_logfdd}
\alias{halfnorm_logfdd}
\title{Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
halfnorm_logfdd(x, v1)

halfnorm_logfdd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Matrix

Matrix
}
\description{
Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol

Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
