% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20c_halfnorm_derivs.R
\name{halfnorm_logfddd}
\alias{halfnorm_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
halfnorm_logfddd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
3d array
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
