% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_loglik}
\alias{halfnorm_loglik}
\title{Log-likelihood function}
\usage{
halfnorm_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Log-likelihood function
}
