% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020b_halfnorm_libs.R
\name{halfnorm_means}
\alias{halfnorm_means}
\title{MLE and RHP predictive means
RHP mean based on the expectation of DMGS equation 2.1}
\usage{
halfnorm_means(means, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 1)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
MLE and RHP predictive means
RHP mean based on the expectation of DMGS equation 2.1
}
