% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_mu2f}
\alias{halfnorm_mu2f}
\title{DMGS equation 3.3, mu2 term}
\usage{
halfnorm_mu2f(alpha, v1, fd1)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, mu2 term
}
