% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_p1f}
\alias{halfnorm_p1f}
\title{DMGS equation 2.1, p1 term}
\usage{
halfnorm_p1f(y, v1, fd1)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Matrix
}
\description{
DMGS equation 2.1, p1 term
}
