% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{ifvectorthenmatrix}
\alias{ifvectorthenmatrix}
\title{If a variable is a vector, convert it to a matrix}
\usage{
ifvectorthenmatrix(t)
}
\arguments{
\item{t}{a vector or matrix of predictors}
}
\value{
Vector
}
\description{
If a variable is a vector, convert it to a matrix
}
