% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101c_invgamma_derivs.R
\name{invgamma_ldda}
\alias{invgamma_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
invgamma_ldda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
The second derivative of the normalized log-likelihood
}
