% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101b_invgamma_libs.R
\name{invgamma_lddd}
\alias{invgamma_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
invgamma_lddd(x, v1, fd1, v2, fd2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood
}
