% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101b_invgamma_libs.R
\name{invgamma_logf}
\alias{invgamma_logf}
\title{Logf for RUST}
\usage{
invgamma_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
