% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101c_invgamma_derivs.R
\name{invgamma_logfdd}
\alias{invgamma_logfdd}
\title{Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
invgamma_logfdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
