% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101b_invgamma_libs.R
\name{invgamma_loglik}
\alias{invgamma_loglik}
\title{log-likelihood function}
\usage{
invgamma_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar
}
\description{
log-likelihood function
}
