% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102c_invgauss_derivs.R
\name{invgauss_lddda}
\alias{invgauss_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
invgauss_lddda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
