% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102b_invgauss_libs.R
\name{invgauss_logf}
\alias{invgauss_logf}
\title{Logf for RUST}
\usage{
invgauss_logf(params, x, prior)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{prior}{logical indicating which prior to use}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
