% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/102b_invgauss_libs.R
\name{invgauss_loglik}
\alias{invgauss_loglik}
\title{log-likelihood function}
\usage{
invgauss_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\description{
log-likelihood function
}
