% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_jeffreys_prior.R
\name{jpf2p}
\alias{jpf2p}
\title{Jeffreys' Prior with two parameters}
\usage{
jpf2p(ggd, detg, ggi)
}
\arguments{
\item{ggd}{gradient of the expected information matrix}

\item{detg}{determinant of the expected information matrix}

\item{ggi}{inverse of the expected information matrix}
}
\description{
Jeffreys' Prior with two parameters
}
