% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_jeffreys_prior.R
\name{jpf3p}
\alias{jpf3p}
\title{Jeffreys' Prior with three parameters}
\usage{
jpf3p(ggd, detg, ggi)
}
\arguments{
\item{ggd}{gradient of the expected information matrix}

\item{detg}{determinant of the expected information matrix}

\item{ggi}{inverse of the expected information matrix}
}
\description{
Jeffreys' Prior with three parameters
}
