% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_jeffreys_prior.R
\name{jpf4p}
\alias{jpf4p}
\title{Jeffreys' Prior with four parameters}
\usage{
jpf4p(ggd, detg, ggi)
}
\arguments{
\item{ggd}{gradient of the expected information matrix}

\item{detg}{determinant of the expected information matrix}

\item{ggi}{inverse of the expected information matrix}
}
\value{
Vector of 4 values
}
\description{
Jeffreys' Prior with four parameters
}
