% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36b_lnorm_dmgs_libs.R
\name{lnorm_dmgs_loglik}
\alias{lnorm_dmgs_loglik}
\title{log-likelihood function}
\usage{
lnorm_dmgs_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\description{
log-likelihood function
}
