% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/036b_lnorm_dmgs_libs.R
\name{lnorm_dmgs_means}
\alias{lnorm_dmgs_means}
\title{MLE and RHP predictive means}
\usage{
lnorm_dmgs_means(means, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 2)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
MLE and RHP predictive means
}
