% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/36b_lnorm_dmgs_libs.R
\name{lnorm_dmgs_mu1f}
\alias{lnorm_dmgs_mu1f}
\title{DMGS equation 3.3, mu1 term}
\usage{
lnorm_dmgs_mu1f(alpha, v1, d1, v2, fd2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 3.3, mu1 term
}
