% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/035c_lnorm_derivs.R
\name{lnorm_f2fa}
\alias{lnorm_f2fa}
\title{The second derivative of the density}
\usage{
lnorm_f2fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density
}
