% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35c_lnorm_derivs.R
\name{lnorm_lddda}
\alias{lnorm_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
lnorm_lddda(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
