% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35b_lnorm_libs.R
\name{lnorm_lmnp}
\alias{lnorm_lmnp}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
lnorm_lmnp(x, v1, d1, v2, fd2, mm, nn, rr)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}

\item{rr}{an index for which derivative to calculate}
}
\description{
One component of the second derivative of the normalized log-likelihood
}
