% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35b_lnorm_libs.R
\name{lnorm_logf}
\alias{lnorm_logf}
\title{Logf for RUST}
\usage{
lnorm_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
