% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061c_lnorm_p1_derivs.R
\name{lnorm_p1_f1fw}
\alias{lnorm_p1_f1fw}
\title{The first derivative of the density for WAIC}
\usage{
lnorm_p1_f1fw(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
The first derivative of the density for WAIC
}
