% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61c_lnorm_p1_derivs.R
\name{lnorm_p1_lddda}
\alias{lnorm_p1_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
lnorm_p1_lddda(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
