% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61b_lnorm_p1_libs.R
\name{lnorm_p1_loglik}
\alias{lnorm_p1_loglik}
\title{Log-normal-with-p1  observed log-likelihood function}
\usage{
lnorm_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar value.
}
\description{
Log-normal-with-p1  observed log-likelihood function
}
