% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061c_lnorm_p1_derivs.R
\name{lnorm_p1_mu2fa}
\alias{lnorm_p1_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
lnorm_p1_mu2fa(alpha, t0, v1, v2, v3)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Matrix
}
\description{
Minus the second derivative of the cdf, at alpha
}
