% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61b_lnorm_p1_libs.R
\name{lnorm_p1_predictordata}
\alias{lnorm_p1_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
lnorm_p1_predictordata(x, t, t0, params)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
