% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35c_lnorm_derivs.R
\name{lnorm_p2fa}
\alias{lnorm_p2fa}
\title{The second derivative of the cdf}
\usage{
lnorm_p2fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
The second derivative of the cdf
}
