% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/35c_lnorm_derivs.R
\name{lnorm_pdd}
\alias{lnorm_pdd}
\title{Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
lnorm_pdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
