% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/035b_lnorm_libs.R
\name{lnorm_waic}
\alias{lnorm_waic}
\title{Waic for RUST}
\usage{
lnorm_waic(waicscores, x, v1hat, v2hat)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}
}
\value{
Two numeric values.
}
\description{
Waic for RUST
}
