% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40b_logis_libs.R
\name{logis_logf}
\alias{logis_logf}
\title{Logf for RUST}
\usage{
logis_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
