% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040b_logis_libs.R
\name{logis_loglik}
\alias{logis_loglik}
\title{log-likelihood function}
\usage{
logis_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
