% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40c_logis_derivs.R
\name{logis_mu2fa}
\alias{logis_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
logis_mu2fa(alpha, v1, v2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
Minus the second derivative of the cdf, at alpha
}
