% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62b_logis_p1_libs.R
\name{logis_p1_logf}
\alias{logis_p1_logf}
\title{Logf for RUST}
\usage{
logis_p1_logf(params, x, t)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\description{
Logf for RUST
}
