% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062b_logis_p1_libs.R
\name{logis_p1_loglik}
\alias{logis_p1_loglik}
\title{Logistic-with-p1  observed log-likelihood function}
\usage{
logis_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar
}
\description{
Logistic-with-p1  observed log-likelihood function
}
