% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62b_logis_p1_libs.R
\name{logis_p1_logscores}
\alias{logis_p1_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
logis_p1_logscores(logscores, x, t, d1, d2, fd3, aderivs = TRUE)
}
\arguments{
\item{logscores}{a logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
