% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062b_logis_p1_libs.R
\name{logis_p1_means}
\alias{logis_p1_means}
\title{Logistic distribution: RHP mean}
\usage{
logis_p1_means(t0, ml_params, lddi, lddd, lambdad_rhp, nx, dim = 2)
}
\arguments{
\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
Logistic distribution: RHP mean
}
