% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062c_logis_p1_derivs.R
\name{logis_p1_p1fa}
\alias{logis_p1_p1fa}
\title{The first derivative of the cdf}
\usage{
logis_p1_p1fa(x, t0, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
The first derivative of the cdf
}
