% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62c_logis_p1_derivs.R
\name{logis_p1_p2fa}
\alias{logis_p1_p2fa}
\title{The second derivative of the cdf}
\usage{
logis_p1_p2fa(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Matrix
}
\description{
The second derivative of the cdf
}
