% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040c_logis_derivs.R
\name{logis_p1fa}
\alias{logis_p1fa}
\title{The first derivative of the cdf}
\usage{
logis_p1fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Vector
}
\description{
The first derivative of the cdf
}
