% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40b_logis_libs.R
\name{logis_p2f}
\alias{logis_p2f}
\title{DMGS equation 3.3, p2 term}
\usage{
logis_p2f(y, v1, d1, v2, fd2)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
3d array
}
\description{
DMGS equation 3.3, p2 term
}
