% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41c_lst_k3_derivs.R
\name{lst_k3_f1fa}
\alias{lst_k3_f1fa}
\title{The first derivative of the density}
\usage{
lst_k3_f1fa(x, v1, v2, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Vector
}
\description{
The first derivative of the density
}
