% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/041c_lst_k3_derivs.R
\name{lst_k3_lddda}
\alias{lst_k3_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
lst_k3_lddda(x, v1, v2, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
3d array
}
\description{
The third derivative of the normalized log-likelihood
}
