% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41b_lst_k3_libs.R
\name{lst_k3_lmnp}
\alias{lst_k3_lmnp}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
lst_k3_lmnp(x, v1, d1, v2, fd2, kdf, mm, nn, rr)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kdf}{the known degrees of freedom parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}

\item{rr}{an index for which derivative to calculate}
}
\description{
One component of the second derivative of the normalized log-likelihood
}
