% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/041b_lst_k3_libs.R
\name{lst_k3_logf}
\alias{lst_k3_logf}
\title{Logf for RUST}
\usage{
lst_k3_logf(params, x, kdf)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
