% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41b_lst_k3_libs.R
\name{lst_k3_loglik}
\alias{lst_k3_loglik}
\title{log-likelihood function}
\usage{
lst_k3_loglik(vv, x, kdf)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{kdf}{the known degrees of freedom parameter}
}
\description{
log-likelihood function
}
