% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41b_lst_k3_libs.R
\name{lst_k3_logscores}
\alias{lst_k3_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
lst_k3_logscores(logscores, x, d1 = 0.01, fd2 = 0.01, kdf, aderivs = TRUE)
}
\arguments{
\item{logscores}{a logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kdf}{the known degrees of freedom parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
