% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41b_lst_k3_libs.R
\name{lst_k3_p2f}
\alias{lst_k3_p2f}
\title{DMGS equation 3.3, p2 term}
\usage{
lst_k3_p2f(y, v1, d1, v2, fd2, kdf)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\description{
DMGS equation 3.3, p2 term
}
