% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063c_lst_p1k3_derivs.R
\name{lst_p1k3_f1fa}
\alias{lst_p1k3_f1fa}
\title{The first derivative of the density for DMGS}
\usage{
lst_p1k3_f1fa(x, t0, v1, v2, v3, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Vector
}
\description{
The first derivative of the density for DMGS
}
