% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063c_lst_p1k3_derivs.R
\name{lst_p1k3_f2fw}
\alias{lst_p1k3_f2fw}
\title{The second derivative of the density for WAIC}
\usage{
lst_p1k3_f2fw(x, t, v1, v2, v3, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density for WAIC
}
