% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063c_lst_p1k3_derivs.R
\name{lst_p1k3_fdd}
\alias{lst_p1k3_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
lst_p1k3_fdd(x, t, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\value{
Matrix
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
