% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63c_lst_p1k3_derivs.R
\name{lst_p1k3_ldda}
\alias{lst_p1k3_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
lst_p1k3_ldda(x, t, v1, v2, v3, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\description{
The second derivative of the normalized log-likelihood
}
