% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63b_lst_p1k3_libs.R
\name{lst_p1k3_logf}
\alias{lst_p1k3_logf}
\title{Logf for RUST}
\usage{
lst_p1k3_logf(params, x, t, kdf)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kdf}{the known degrees of freedom parameter}
}
\description{
Logf for RUST
}
