% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063b_lst_p1k3_libs.R
\name{lst_p1k3_loglik}
\alias{lst_p1k3_loglik}
\title{LST-with-p1  observed log-likelihood function}
\usage{
lst_p1k3_loglik(vv, x, t, kdf)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Scalar
}
\description{
LST-with-p1  observed log-likelihood function
}
