% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063b_lst_p1k3_libs.R
\name{lst_p1k3_logscores}
\alias{lst_p1k3_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
lst_p1k3_logscores(logscores, x, t, d1, d2, fd3, kdf, aderivs)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kdf}{the known degrees of freedom parameter}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
