% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063b_lst_p1k3_libs.R
\name{lst_p1k3_mu1f}
\alias{lst_p1k3_mu1f}
\title{DMGS equation 3.3, mu1 term}
\usage{
lst_p1k3_mu1f(alpha, t0, v1, d1, v2, d2, v3, fd3, kdf)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Matrix
}
\description{
DMGS equation 3.3, mu1 term
}
