% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63b_lst_p1k3_libs.R
\name{lst_p1k3_p1f}
\alias{lst_p1k3_p1f}
\title{DMGS equation 2.1, p1 term}
\usage{
lst_p1k3_p1f(y, t0, v1, d1, v2, d2, v3, fd3, kdf)
}
\arguments{
\item{y}{value of random variable}

\item{t0}{value of predictor}

\item{v1}{first parameter}

\item{d1}{delta for numerical derivative}

\item{v2}{second parameter}

\item{d2}{delta for numerical derivative}

\item{v3}{third parameter}

\item{fd3}{fractional delta for numerical derivative}

\item{kdf}{the known number of degrees of freedom}
}
\description{
DMGS equation 2.1, p1 term
}
