% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63b_lst_p1k3_libs.R
\name{lst_p1k3_predictordata}
\alias{lst_p1k3_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
lst_p1k3_predictordata(predictordata, x, t, t0, params, kdf)
}
\arguments{
\item{predictordata}{a logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}

\item{kdf}{the known degrees of freedom parameter}
}
\description{
Predicted Parameter and Generalized Residuals
}
