% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63b_lst_p1k3_libs.R
\name{lst_p1k3_setics}
\alias{lst_p1k3_setics}
\title{Set initial conditions}
\usage{
lst_p1k3_setics(x, t, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{ics}{initial conditions for the maximum likelihood search}
}
\value{
Vector
}
\description{
Set initial conditions
}
