% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{make_maic}
\alias{make_maic}
\title{Calculate MAIC}
\usage{
make_maic(ml_value, nparams)
}
\arguments{
\item{ml_value}{maximum of the likelihood}

\item{nparams}{number of parameters}
}
\value{
Vector of 3 values
Returns the two compoments of MAIC, and their sum
}
\description{
Calculate MAIC
}
