% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{make_waic}
\alias{make_waic}
\title{Make WAIC}
\usage{
make_waic(x, fhatx, lddi, lddd, f1f, lambdad, f2f, dim)
}
\arguments{
\item{x}{the training data}

\item{fhatx}{density of x at the maximum likelihood parameters}

\item{lddi}{inverse of the second derivative log-likelihood matrix}

\item{lddd}{the third derivative log-likelihood tensor}

\item{f1f}{the f1 term from DMGS equation 2.1}

\item{lambdad}{the slope of the log prior}

\item{f2f}{the f2 term from DMGS equation 2.1}

\item{dim}{number of free parameters}
}
\value{
Two scalars
}
\description{
Make WAIC
}
