% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{makebetat0}
\alias{makebetat0}
\title{Calculate the location parameter when there are predictors (single time point)}
\usage{
makebetat0(nt, params, t0)
}
\arguments{
\item{nt}{the number of columns in \code{t}}

\item{params}{model parameters for calculating logf}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}
}
\value{
Vector
}
\description{
Calculate the location parameter when there are predictors (single time point)
}
