% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{makebetatm}
\alias{makebetatm}
\title{Calculate the location parameter when there are predictors (multiple time points)}
\usage{
makebetatm(nt, params, t)
}
\arguments{
\item{nt}{the number of columns in \code{t}}

\item{params}{model parameters for calculating logf}

\item{t}{a vector or matrix of predictors}
}
\value{
Vector
}
\description{
Calculate the location parameter when there are predictors (multiple time points)
}
