% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{makemuhat0}
\alias{makemuhat0}
\title{Make muhat0}
\usage{
makemuhat0(t0, n0, t, mle_params)
}
\arguments{
\item{t0}{the value of the predictor vector at which to make the prediction (if n0 not specified)}

\item{n0}{the position in the predictor vector at which to make the prediction (positive integer less than or equal to the length of \eqn{x}) (if t0 not specified)}

\item{t}{predictor}

\item{mle_params}{MLE params}
}
\value{
Scalar
}
\description{
Make muhat0
}
