% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{makeq}
\alias{makeq}
\title{Calculates quantiles from simulations by inverting the Hazen CDF}
\usage{
makeq(yy, pp)
}
\arguments{
\item{yy}{vector of samples}

\item{pp}{vector of probabilities}
}
\value{
Vector
}
\description{
Calculates quantiles from simulations by inverting the Hazen CDF
}
