% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{maket0}
\alias{maket0}
\title{Determine t0}
\usage{
maket0(t0, n0, t)
}
\arguments{
\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{n0}{an index for the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{t}{a vector or matrix of predictors}
}
\value{
Scalar
}
\description{
Determine t0
}
