% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{maketresid0}
\alias{maketresid0}
\title{Make ta0}
\usage{
maketresid0(t0, n0, t)
}
\arguments{
\item{t0}{the value of the predictor vector at which to make the prediction (if n0 not specified)}

\item{n0}{the position in the predictor vector at which to make the prediction (positive integer less than or equal to the length of \eqn{x}) (if t0 not specified)}

\item{t}{predictor}
}
\value{
Scalar
}
\description{
Make ta0
}
