% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007_manf.R
\name{manf}
\alias{manf}
\title{Blank function I use for setting up the man page information for the functions}
\usage{
manf(
  dim,
  vv,
  ml_params,
  nx,
  nxx,
  x,
  xx,
  t,
  t1,
  t2,
  t3,
  tt,
  tt1,
  tt2,
  tt3,
  tt2d,
  tt3d,
  t0,
  t01,
  t02,
  t03,
  t10,
  t20,
  t30,
  n0,
  n10,
  n20,
  p,
  n,
  y,
  ics,
  ta,
  ta0,
  muhat0,
  v1,
  v1hat,
  v1h,
  d1,
  fd1,
  v2,
  v2hat,
  v2h,
  d2,
  fd2,
  v3,
  v3hat,
  v3h,
  d3,
  fd3,
  v4,
  v4hat,
  v4h,
  d4,
  fd4,
  v5,
  v5hat,
  v5h,
  d5,
  v6,
  v6hat,
  v6h,
  d6,
  minxi,
  maxxi,
  ximin,
  ximax,
  fdalpha,
  kscale,
  kloc,
  kshape,
  kdf,
  kbeta,
  alpha,
  ymn,
  slope,
  mu,
  sigma,
  sigma1,
  sigma2,
  scale,
  shape,
  xi,
  xi1,
  xi2,
  lambda,
  log,
  mm,
  nn,
  rr,
  lddi,
  lddi_k2,
  lddi_k3,
  lddi_k4,
  lddd,
  lddd_k2,
  lddd_k3,
  lddd_k4,
  lambdad,
  lambdad_cp,
  lambdad_rhp,
  lambdad_flat,
  lambdad_rh_mle,
  lambdad_rh_flat,
  lambdad_jp,
  lambdad_custom,
  means,
  waicscores,
  logscores,
  extramodels,
  pdf,
  predictordata,
  nonnegslopesonly,
  rnonnegslopesonly,
  customprior,
  prior,
  params,
  yy,
  pp,
  dlogpi,
  debug,
  centering,
  aderivs
)
}
\arguments{
\item{dim}{number of parameters}

\item{vv}{parameters}

\item{ml_params}{parameters}

\item{nx}{length of training data}

\item{nxx}{length of training data}

\item{x}{a vector of training data values}

\item{xx}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{tt}{a vector of predictors}

\item{tt1}{a vector of predictors for the mean}

\item{tt2}{a vector of predictors for the sd}

\item{tt3}{a vector of predictors for the shape}

\item{tt2d}{a matrix of predictors (nx by 2)}

\item{tt3d}{a matrix of predictors (nx by 3)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{t03}{a single value of the predictor (specify either \code{t03} or \code{n03} but not both)}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{t30}{a single value of the predictor for the shape (specify either \code{t30} or \code{n30} but not both)}

\item{n0}{an index for the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{n10}{an index for the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{n20}{an index for the predictor for the sd (specify either \code{t10} or \code{n10} but not both)}

\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{n}{number of random samples required}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{ics}{initial conditions for the maximum likelihood search}

\item{ta}{predictor residuals}

\item{ta0}{predictor residual at the point being predicted}

\item{muhat0}{muhat at the point being predicted}

\item{v1}{first parameter}

\item{v1hat}{first parameter}

\item{v1h}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{v2hat}{second parameter}

\item{v2h}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{v3hat}{third parameter}

\item{v3h}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd3}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{v4hat}{fourth parameter}

\item{v4h}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{fd4}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v5}{fifth parameter}

\item{v5hat}{fifth parameter}

\item{v5h}{fifth parameter}

\item{d5}{the delta used in the numerical derivatives with respect to the parameter}

\item{v6}{sixth parameter}

\item{v6hat}{sixth parameter}

\item{v6h}{sixth parameter}

\item{d6}{the delta used in the numerical derivatives with respect to the parameter}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{ximin}{minimum value of shape parameter xi}

\item{ximax}{maximum value of shape parameter xi}

\item{fdalpha}{the fractional delta used in the numerical derivatives with respect to probability, for calculating the pdf as a function of quantiles}

\item{kscale}{the known scale parameter}

\item{kloc}{the known location parameter}

\item{kshape}{the known shape parameter}

\item{kdf}{the known degrees of freedom parameter}

\item{kbeta}{the known beta parameter}

\item{alpha}{a vector of values of alpha (one minus probability)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{mu}{the location parameter of the distribution}

\item{sigma}{the sigma parameter of the distribution}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}

\item{scale}{the scale parameter of the distribution}

\item{shape}{the shape parameter of the distribution}

\item{xi}{the shape parameter of the distribution}

\item{xi1}{first coefficient for the shape parameter of the distribution}

\item{xi2}{second coefficient for the shape parameter of the distribution}

\item{lambda}{the lambda parameter of the distribution}

\item{log}{logical for the density evaluation}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}

\item{rr}{an index for which derivative to calculate}

\item{lddi}{inverse observed information matrix}

\item{lddi_k2}{inverse observed information matrix, fixed shape parameter}

\item{lddi_k3}{inverse observed information matrix, fixed shape parameter}

\item{lddi_k4}{inverse observed information matrix, fixed shape parameter}

\item{lddd}{third derivative of log-likelihood}

\item{lddd_k2}{third derivative of log-likelihood, fixed shape parameter}

\item{lddd_k3}{third derivative of log-likelihood, fixed shape parameter}

\item{lddd_k4}{third derivative of log-likelihood, fixed shape parameter}

\item{lambdad}{derivative of the log prior}

\item{lambdad_cp}{derivative of the log prior}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{lambdad_flat}{derivative of the log flat prior}

\item{lambdad_rh_mle}{derivative of the log CRHP-MLE prior}

\item{lambdad_rh_flat}{derivative of the log CRHP-FLAT prior}

\item{lambdad_jp}{derivative of the log JP prior}

\item{lambdad_custom}{custom value of the derivative of the log prior}

\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{extramodels}{logical that indicates whether to add three additional prediction models}

\item{pdf}{logical that indicates whether to return density functions evaluated at quantiles specified by input probabilities}

\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{nonnegslopesonly}{logical that indicates whether to disallow non-negative slopes}

\item{rnonnegslopesonly}{logical that indicates whether to disallow non-negative slopes}

\item{customprior}{a custom value for the slope of the log prior at the maxlik estimate}

\item{prior}{logical indicating which prior to use}

\item{params}{model parameters for calculating logf}

\item{yy}{vector of samples}

\item{pp}{vector of probabilities}

\item{dlogpi}{gradient of the log prior}

\item{debug}{debug flag}

\item{centering}{indicates whether the routine should center the data or not}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
No return value
}
\description{
Blank function I use for setting up the man page information for the functions
}
