% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_modelselection_flat.R
\name{modelselection_flat}
\alias{modelselection_flat}
\title{Illustration of Model Selection Among 18 Distributions from the \code{fitdistcp} Package}
\usage{
modelselection_flat(x, evd = FALSE)
}
\arguments{
\item{x}{data vector}

\item{evd}{a flag to include GEVD and GPD from the analysis}
}
\value{
Plots QQ plots to the screen, for each of the models,
and returns a data frame containing
\itemize{
\item AIC scores (times -0.5), AIC weights
\item  WAIC1 scores, WAIC1 weights
\item WAIC2 scores, WAIC2 weights
\item logscores, logscore weights
\item maximum likelihood and calibrating prior means
\item maximum likelihood and calibrating prior standard deviations
}
}
\description{
Applies model selection using AIC, WAIC1, WAIC2 and leave-one-out logscore
to the input data \eqn{x},
for 18 models in the \code{fitdistcp} packages
(although for the GEV and GPD, the logscore is NA).

The code is straightforward, and the point is to illustrate what is
possible using the outputs from the \code{fitdistcp} routines.

The input data is automatically shifted so that the minimum value is positive,
to avoid errors from the models that require positive random variable.

For the Pareto, the data is further shifted so that the minimum value is slightly greater than 1.

GEVD and GPD are not included by default, because they are temperamental in that
they do not work if the shape parameter is extreme.
}
\details{
The 18 models are:
\code{exp},
\code{pareto_k2},
\code{halfnorm},
\code{norm},
\code{lnorm},
\code{gnorm_k3},
\code{gumbel},
\code{frechet_k1},
\code{weibull},
\code{gev_k3},
\code{logis},
\code{lst_k3},
\code{cauchy},
\code{gamma},
\code{invgamma},
\code{invgauss},
\code{gev} and
\code{gpd_k1}.
}
\examples{
set.seed(1)
nx=50
x=rlnorm(nx)
print(modelselection_flat(x))


}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
