% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{movexiawayfromzero}
\alias{movexiawayfromzero}
\title{Move xi away from zero a bit}
\usage{
movexiawayfromzero(xi)
}
\arguments{
\item{xi}{xi}
}
\value{
Scalar
}
\description{
Move xi away from zero a bit
}
