% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{nopdfcdfmsg}
\alias{nopdfcdfmsg}
\title{Message to explain why GEV and GPD d*** and p*** routines don't return DMGS pdfs and cdfs}
\usage{
nopdfcdfmsg(yy, pp)
}
\arguments{
\item{yy}{vector of samples}

\item{pp}{vector of probabilities}
}
\description{
Message to explain why GEV and GPD d*** and p*** routines don't return DMGS pdfs and cdfs
}
