% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030b_norm_libs.R
\name{norm_boot}
\alias{norm_boot}
\title{Bootstrap}
\usage{
norm_boot(x, n)
}
\arguments{
\item{x}{a vector of training data values}

\item{n}{number of random samples required}
}
\value{
A list containing a matrix of simulated parameter values
}
\description{
Bootstrap
}
