% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031b_norm_dmgs_libs.R
\name{norm_dmgs_loglik}
\alias{norm_dmgs_loglik}
\title{log-likelihood function}
\usage{
norm_dmgs_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
log-likelihood function
}
