% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31b_norm_dmgs_libs.R
\name{norm_dmgs_p1f}
\alias{norm_dmgs_p1f}
\title{DMGS equation 3.3, p1 term}
\usage{
norm_dmgs_p1f(y, v1, d1, v2, fd2)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\value{
Matrix
}
\description{
DMGS equation 3.3, p1 term
}
