% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030c_norm_derivs.R
\name{norm_f2fa}
\alias{norm_f2fa}
\title{The second derivative of the density}
\usage{
norm_f2fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
The second derivative of the density
}
