% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30b_norm_libs.R
\name{norm_gmn}
\alias{norm_gmn}
\title{One component of the second derivative of the expected log-likelihood}
\usage{
norm_gmn(alpha, v1, d1, v2, fd2, mm, nn)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the second derivative of the expected log-likelihood
}
