% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30b_norm_libs.R
\name{norm_lddd}
\alias{norm_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
norm_lddd(x, v1, d1, v2, fd2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
Third derivative tensor of the normalized log-likelihood
}
