% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030b_norm_libs.R, R/031b_norm_dmgs_libs.R
\name{norm_lmnp}
\alias{norm_lmnp}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
norm_lmnp(x, v1, d1, v2, fd2, mm, nn, rr)

norm_lmnp(x, v1, d1, v2, fd2, mm, nn, rr)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}

\item{rr}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the second derivative of the normalized log-likelihood

One component of the third derivative of the normalized log-likelihood
}
