% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030b_norm_libs.R
\name{norm_logf}
\alias{norm_logf}
\title{Logf for RUST}
\usage{
norm_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
