% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030b_norm_libs.R
\name{norm_ml_params}
\alias{norm_ml_params}
\title{Maximum likelihood estimator}
\usage{
norm_ml_params(x)
}
\arguments{
\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Maximum likelihood estimator
}
