% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030c_norm_derivs.R
\name{norm_mu1fa}
\alias{norm_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
norm_mu1fa(alpha, v1, v2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Vector
}
\description{
Minus the first derivative of the cdf, at alpha
}
