% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30c_norm_derivs.R
\name{norm_mu2fa}
\alias{norm_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
norm_mu2fa(alpha, v1, v2)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Minus the second derivative of the cdf, at alpha
}
