% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_exampledata}
\alias{norm_p12_exampledata}
\title{Norm_p12 example data}
\usage{
norm_p12_exampledata(iseed)
}
\arguments{
\item{iseed}{The random seed}
}
\value{
A list containing data to run an example
}
\description{
Norm_p12 example data
}
