% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080c_norm_p12_derivs.R
\name{norm_p12_f1fa}
\alias{norm_p12_f1fa}
\title{The first derivative of the density}
\usage{
norm_p12_f1fa(x, t1, t2, v1, v2, v3, v4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\description{
The first derivative of the density
}
