% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_ldd}
\alias{norm_p12_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
norm_p12_ldd(x, t1, t2, v1, d1, v2, d2, v3, d3, v4, d4)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}
}
\value{
Matrix
}
\description{
Second derivative matrix of the normalized log-likelihood
}
