% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_logf}
\alias{norm_p12_logf}
\title{Logf for RUST}
\usage{
norm_p12_logf(params, x, t1, t2, nonnegslopesonly = FALSE)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{nonnegslopesonly}{logical that indicates whether to disallow non-negative slopes}
}
\description{
Logf for RUST
}
