% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_loglik}
\alias{norm_p12_loglik}
\title{observed log-likelihood function}
\usage{
norm_p12_loglik(vv, x, t1, t2)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}
}
\description{
observed log-likelihood function
}
