% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080c_norm_p12_derivs.R
\name{norm_p12_mu2fa}
\alias{norm_p12_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
norm_p12_mu2fa(alpha, t1, t2, v1, v2, v3, v4)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}

\item{v4}{fourth parameter}
}
\description{
Minus the second derivative of the cdf, at alpha
}
