% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_p1f}
\alias{norm_p12_p1f}
\title{DMGS equation 2.1, p1 term}
\usage{
norm_p12_p1f(y, t10, t20, v1, d1, v2, d2, v3, d3, v4, d4)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{v4}{fourth parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}
}
\description{
DMGS equation 2.1, p1 term
}
