% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_predictordata}
\alias{norm_p12_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
norm_p12_predictordata(predictordata, x, t1, t2, t10, t20, params)
}
\arguments{
\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{params}{model parameters for calculating logf}
}
\description{
Predicted Parameter and Generalized Residuals
}
