% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60c_norm_p1_derivs.R,
%   R/60c_norm_p1_derivs_old.R
\name{norm_p1_f1fa}
\alias{norm_p1_f1fa}
\title{The first derivative of the density}
\usage{
norm_p1_f1fa(x, t, v1, v2, v3)

norm_p1_f1fa(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
The first derivative of the density

The first derivative of the density
}
