% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060c_norm_p1_derivs.R,
%   R/060c_norm_p1_derivs_old.R
\name{norm_p1_fd}
\alias{norm_p1_fd}
\title{First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
norm_p1_fd(x, t, v1, v2, v3)

norm_p1_fd(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Vector
}
\description{
First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol

First derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
