% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60c_norm_p1_derivs_old.R
\name{norm_p1_lddda}
\alias{norm_p1_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
norm_p1_lddda(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
The third derivative of the normalized log-likelihood
}
