% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60b_norm_p1_libs.R
\name{norm_p1_logf}
\alias{norm_p1_logf}
\title{Logf for RUST}
\usage{
norm_p1_logf(params, x, t)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\description{
Logf for RUST
}
