% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060c_norm_p1_derivs.R,
%   R/060c_norm_p1_derivs_old.R
\name{norm_p1_logfdd}
\alias{norm_p1_logfdd}
\title{Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
norm_p1_logfdd(x, t, v1, v2, v3)

norm_p1_logfdd(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Matrix
}
\description{
Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol

Second derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
