% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60c_norm_p1_derivs_old.R
\name{norm_p1_logfddd}
\alias{norm_p1_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
norm_p1_logfddd(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
