% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60b_norm_p1_libs.R
\name{norm_p1_loglik}
\alias{norm_p1_loglik}
\title{Normal-with-p1  observed log-likelihood function}
\usage{
norm_p1_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\description{
Normal-with-p1  observed log-likelihood function
}
