% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60b_norm_p1_libs.R
\name{norm_p1_logscores}
\alias{norm_p1_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
norm_p1_logscores(logscores, x, t, aderivs = TRUE)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
