% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060b_norm_p1_libs.R
\name{norm_p1_mlparams}
\alias{norm_p1_mlparams}
\title{Maximum likelihood estimator}
\usage{
norm_p1_mlparams(x, t)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Vector
}
\description{
Maximum likelihood estimator
}
