% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30c_norm_derivs.R
\name{norm_p1fa}
\alias{norm_p1fa}
\title{The first derivative of the cdf}
\usage{
norm_p1fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The first derivative of the cdf
}
