% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30b_norm_libs.R
\name{norm_unbiasedv_params}
\alias{norm_unbiasedv_params}
\title{Method of moments estimator}
\usage{
norm_unbiasedv_params(x)
}
\arguments{
\item{x}{a vector of training data values}
}
\value{
Vector
}
\description{
Method of moments estimator
}
