% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030b_norm_libs.R
\name{norm_waic}
\alias{norm_waic}
\title{Waic}
\usage{
norm_waic(waicscores, x, v1hat, v2hat)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}
}
\value{
Two numeric values.
}
\description{
Waic
}
