% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11c_pareto_k2_derivs.R
\name{pareto_k2_f1fa}
\alias{pareto_k2_f1fa}
\title{The first derivative of the density}
\usage{
pareto_k2_f1fa(x, v1, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
Vector
}
\description{
The first derivative of the density
}
