% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11b_pareto_k2_libs.R
\name{pareto_k2_f2f}
\alias{pareto_k2_f2f}
\title{DMGS equation 2.1, f2 term}
\usage{
pareto_k2_f2f(y, v1, fd1, kscale)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kscale}{the known scale parameter}
}
\value{
3d array
}
\description{
DMGS equation 2.1, f2 term
}
