% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011c_pareto_k2_derivs.R,
%   R/11c_pareto_k2_derivs.R
\name{pareto_k2_f2fa}
\alias{pareto_k2_f2fa}
\title{The second derivative of the density}
\usage{
pareto_k2_f2fa(x, v1, kscale)

pareto_k2_f2fa(x, v1, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
Matrix

Matrix
}
\description{
The second derivative of the density

The second derivative of the density
}
