% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11c_pareto_k2_derivs.R
\name{pareto_k2_ldda}
\alias{pareto_k2_ldda}
\title{The second derivative of the normalized log-likelihood}
\usage{
pareto_k2_ldda(x, v1, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
Matrix
}
\description{
The second derivative of the normalized log-likelihood
}
