% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11b_pareto_k2_libs.R
\name{pareto_k2_lddd}
\alias{pareto_k2_lddd}
\title{Third derivative tensor of the log-likelihood}
\usage{
pareto_k2_lddd(x, v1, fd1, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{kscale}{the known scale parameter}
}
\description{
Third derivative tensor of the log-likelihood
}
