% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011c_pareto_k2_derivs.R,
%   R/11c_pareto_k2_derivs.R
\name{pareto_k2_lddda}
\alias{pareto_k2_lddda}
\title{The third derivative of the normalized log-likelihood}
\usage{
pareto_k2_lddda(x, v1, kscale)

pareto_k2_lddda(x, v1, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
3d array

3d array
}
\description{
The third derivative of the normalized log-likelihood

The third derivative of the normalized log-likelihood
}
