% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011b_pareto_k2_libs.R
\name{pareto_k2_logf}
\alias{pareto_k2_logf}
\title{Logf for RUST}
\usage{
pareto_k2_logf(params, x, kscale)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{kscale}{the known scale parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
