% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011c_pareto_k2_derivs.R,
%   R/11c_pareto_k2_derivs.R
\name{pareto_k2_logfddd}
\alias{pareto_k2_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
pareto_k2_logfddd(x, v1, v2)

pareto_k2_logfddd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
3d array

3d array
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol

Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
