% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011b_pareto_k2_libs.R
\name{pareto_k2_logscores}
\alias{pareto_k2_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
pareto_k2_logscores(logscores, x, kscale)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{kscale}{the known scale parameter}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
