% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011b_pareto_k2_libs.R
\name{pareto_k2_ml_params}
\alias{pareto_k2_ml_params}
\title{Maximum likelihood estimator}
\usage{
pareto_k2_ml_params(x, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{kscale}{the known scale parameter}
}
\value{
Scalar value.
}
\description{
Maximum likelihood estimator
}
