% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11c_pareto_k2_derivs.R
\name{pareto_k2_pdd}
\alias{pareto_k2_pdd}
\title{Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
pareto_k2_pdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Matrix
}
\description{
Second derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
