% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011b_pareto_k2_libs.R
\name{pareto_k2_waic}
\alias{pareto_k2_waic}
\title{Waic}
\usage{
pareto_k2_waic(waicscores, x, v1hat, kscale)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{v1hat}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
Two numeric values.
}
\description{
Waic
}
