% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{pareto_p1k2_f2f}
\alias{pareto_p1k2_f2f}
\title{DMGS equation 2.1, f2 term}
\usage{
pareto_p1k2_f2f(y, t0, v1, d1, v2, d2, kscale)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kscale}{the known scale parameter}
}
\value{
3d array
}
\description{
DMGS equation 2.1, f2 term
}
