% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{pareto_p1k2_ldd}
\alias{pareto_p1k2_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
pareto_p1k2_ldd(x, t, v1, d1, v2, d2, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kscale}{the known scale parameter}
}
\value{
Square scalar matrix
}
\description{
Second derivative matrix of the normalized log-likelihood
}
