% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56b_pareto_p1k2_libs.R
\name{pareto_p1k2_lddd}
\alias{pareto_p1k2_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
pareto_p1k2_lddd(x, t, v1, d1, v2, d2, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{kscale}{the known scale parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood
}
