% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56b_pareto_p1k2_libs.R
\name{pareto_p1k2_logf}
\alias{pareto_p1k2_logf}
\title{Logf for RUST}
\usage{
pareto_p1k2_logf(params, x, t, kscale)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kscale}{the known scale parameter}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
