% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{pareto_p1k2_loglik}
\alias{pareto_p1k2_loglik}
\title{observed log-likelihood function}
\usage{
pareto_p1k2_loglik(vv, x, t, kscale)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kscale}{the known scale parameter}
}
\value{
Scalar value.
}
\description{
observed log-likelihood function
}
