% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{pareto_p1k2_logscores}
\alias{pareto_p1k2_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
pareto_p1k2_logscores(logscores, x, t, kscale, debug)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{kscale}{the known scale parameter}

\item{debug}{debug flag}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
