% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56c_pareto_p1k2_derivs.R
\name{pareto_p1k2_mu1fa}
\alias{pareto_p1k2_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
pareto_p1k2_mu1fa(alpha, t, v1, v2, kscale)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kscale}{the known scale parameter}
}
\description{
Minus the first derivative of the cdf, at alpha
}
