% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056c_pareto_p1k2_derivs.R
\name{pareto_p1k2_mu2fa}
\alias{pareto_p1k2_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
pareto_p1k2_mu2fa(alpha, t0, v1, v2, kscale)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kscale}{the known scale parameter}
}
\value{
Matrix
}
\description{
Minus the second derivative of the cdf, at alpha
}
