% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56c_pareto_p1k2_derivs.R
\name{pareto_p1k2_p1fa}
\alias{pareto_p1k2_p1fa}
\title{The first derivative of the cdf}
\usage{
pareto_p1k2_p1fa(x, t, v1, v2, kscale)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{kscale}{the known scale parameter}
}
\description{
The first derivative of the cdf
}
