% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056b_pareto_p1k2_libs.R
\name{pareto_p1k2_predictordata}
\alias{pareto_p1k2_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
pareto_p1k2_predictordata(predictordata, x, t, t0, params, kscale)
}
\arguments{
\item{predictordata}{logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}

\item{kscale}{the known scale parameter}
}
\value{
Two vectors
}
\description{
Predicted Parameter and Generalized Residuals
}
