% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/64b_cauchy_p1_libs.R
\name{pcauchy_p1}
\alias{pcauchy_p1}
\title{Cauchy-with-p1 distribution function}
\usage{
pcauchy_p1(x, t0, ymn, slope, scale)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{scale}{the scale parameter of the distribution}
}
\description{
Cauchy-with-p1 distribution function
}
