% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{pexp_p1}
\alias{pexp_p1}
\title{Exponential-with-p1 distribution function}
\usage{
pexp_p1(x, t0, ymn, slope)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}
}
\value{
Vector
}
\description{
Exponential-with-p1 distribution function
}
