% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{pgev_p1}
\alias{pgev_p1}
\title{GEVD-with-p1: Distribution function}
\usage{
pgev_p1(y, t0, ymn, slope, sigma, xi)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}

\item{xi}{the shape parameter of the distribution}
}
\description{
GEVD-with-p1: Distribution function
}
