% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/153b_gev_p1n_libs.R
\name{pgev_p1n}
\alias{pgev_p1n}
\title{GEVD-with-p1: Distribution function}
\usage{
pgev_p1n(y, t0, params)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Vector
}
\description{
GEVD-with-p1: Distribution function
}
