% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061b_lnorm_p1_libs.R
\name{plnorm_p1}
\alias{plnorm_p1}
\title{Normal-with-p1 distribution function}
\usage{
plnorm_p1(x, t0, ymn, slope, sigma)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}
}
\value{
Vector
}
\description{
Normal-with-p1 distribution function
}
