% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063b_lst_p1k3_libs.R
\name{plst_p1k3}
\alias{plst_p1k3}
\title{LST-with-p1 distribution function}
\usage{
plst_p1k3(x, t0, ymn, slope, sigma, kdf)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}

\item{kdf}{the known degrees of freedom parameter}
}
\value{
Vector
}
\description{
LST-with-p1 distribution function
}
