% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{pnorm_p12}
\alias{pnorm_p12}
\title{Normal-with-p12: Distribution function}
\usage{
pnorm_p12(y, t10, t20, ymn, slope, sigma1, sigma2)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}
}
\value{
Vector
}
\description{
Normal-with-p12: Distribution function
}
