% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060b_norm_p1_libs.R
\name{pnorm_p1_formula}
\alias{pnorm_p1_formula}
\title{Linear regression formula, densities}
\usage{
pnorm_p1_formula(y, tresid, tresid0, nx, muhat0, v3hat)
}
\arguments{
\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{tresid}{predictor residuals}

\item{tresid0}{predictor residual at the point being predicted}

\item{nx}{length of training data}

\item{muhat0}{muhat at the point being predicted}

\item{v3hat}{third parameter}
}
\value{
Vector
}
\description{
Linear regression formula, densities
}
