% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25b_unif_libs.R
\name{punif_formula}
\alias{punif_formula}
\title{Predictive CDFs}
\usage{
punif_formula(x, y)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}
}
\description{
Predictive CDFs
}
