% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55b_exp_p1_libs.R
\name{qexp_p1}
\alias{qexp_p1}
\title{-with-p1 quantile function}
\usage{
qexp_p1(p, t0, ymn, slope)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}
}
\value{
Vector
}
\description{
-with-p1 quantile function
}
