% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71b_frechet_p2k1_libs.R
\name{qfrechet_p2k1}
\alias{qfrechet_p2k1}
\title{Frechet_k1-with-p2 quantile function}
\usage{
qfrechet_p2k1(p, t0, ymn, slope, lambda, kloc)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{lambda}{the lambda parameter of the distribution}

\item{kloc}{the known location parameter}
}
\description{
Frechet_k1-with-p2 quantile function
}
