% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/151b_gev_p12_libs.R
\name{qgev_p12}
\alias{qgev_p12}
\title{GEVD-with-p1: Quantile function}
\usage{
qgev_p12(p, t1, t2, ymn, slope, sigma1, sigma2, xi)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}

\item{xi}{the shape parameter of the distribution}
}
\description{
GEVD-with-p1: Quantile function
}
