% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{qgev_p123}
\alias{qgev_p123}
\title{GEVD-with-p1: Quantile function}
\usage{
qgev_p123(p, t1, t2, t3, ymn, slope, sigma1, sigma2, xi1, xi2)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}

\item{xi1}{first coefficient for the shape parameter of the distribution}

\item{xi2}{second coefficient for the shape parameter of the distribution}
}
\description{
GEVD-with-p1: Quantile function
}
