% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_ppm_temporary.R
\name{qgev_p1_ppm}
\alias{qgev_p1_ppm}
\title{Temporary dummy for one of the ppm models}
\usage{
qgev_p1_ppm(x, t, n0, p)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector of predictors, such that \code{length(t)=length(x)}}

\item{n0}{an index for the predictor
(specify either \code{t0} or \code{n0} but not both)}

\item{p}{a vector of probabilities at which to generate predictive quantiles}
}
\description{
Temporary dummy for one of the ppm models
}
