% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1n_libs.R
\name{qgev_p1n}
\alias{qgev_p1n}
\title{GEVD-with-p1: Quantile function}
\usage{
qgev_p1n(p, t0, params)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}
}
\value{
Vector
}
\description{
GEVD-with-p1: Quantile function
}
