% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_ppm_temporary.R
\name{qgpd_k1_ppm}
\alias{qgpd_k1_ppm}
\title{Temporary dummy for one of the ppm models}
\usage{
qgpd_k1_ppm(x, p)
}
\arguments{
\item{x}{a vector of training data values}

\item{p}{a vector of probabilities at which to generate predictive quantiles}
}
\description{
Temporary dummy for one of the ppm models
}
