% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61b_lnorm_p1_libs.R
\name{qlnorm_p1}
\alias{qlnorm_p1}
\title{Normal-with-p1 quantile function}
\usage{
qlnorm_p1(p, t0, ymn, slope, sigma)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma}{the sigma parameter of the distribution}
}
\description{
Normal-with-p1 quantile function
}
