% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/62b_logis_p1_libs.R
\name{qlogis_p1}
\alias{qlogis_p1}
\title{Logistic-with-p1 quantile function}
\usage{
qlogis_p1(p, t0, ymn, slope, scale)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{scale}{the scale parameter of the distribution}
}
\description{
Logistic-with-p1 quantile function
}
