% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{qnorm_p12}
\alias{qnorm_p12}
\title{Normal-with-p12: Quantile function}
\usage{
qnorm_p12(p, t10, t20, ymn, slope, sigma1, sigma2)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}
}
\description{
Normal-with-p12: Quantile function
}
