% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060b_norm_p1_libs.R
\name{qnorm_p1_formula}
\alias{qnorm_p1_formula}
\title{Linear regression formula, quantiles}
\usage{
qnorm_p1_formula(alpha, ta, ta0, nx, muhat0, v3hat)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{ta}{predictor residuals}

\item{ta0}{predictor residual at the point being predicted}

\item{nx}{length of training data}

\item{muhat0}{muhat at the point being predicted}

\item{v3hat}{third parameter}
}
\value{
Vector
}
\description{
Linear regression formula, quantiles
}
