% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/56b_pareto_p1k2_libs.R
\name{qpareto_p1k2}
\alias{qpareto_p1k2}
\title{pareto_k1-with-p2 quantile function}
\usage{
qpareto_p1k2(p, t0, ymn, slope, kscale)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{kscale}{the known scale parameter}
}
\description{
pareto_k1-with-p2 quantile function
}
