% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/25b_unif_libs.R
\name{qunif_formula}
\alias{qunif_formula}
\title{Predictive Quantiles}
\usage{
qunif_formula(x, p)
}
\arguments{
\item{x}{a vector of training data values}

\item{p}{a vector of probabilities at which to generate predictive quantiles}
}
\description{
Predictive Quantiles
}
