% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073b_weibull_p2_libs.R
\name{qweibull_p2}
\alias{qweibull_p2}
\title{Weibull-with-p1 quantile function}
\usage{
qweibull_p2(p, t0, shape, ymn, slope)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{shape}{the shape parameter of the distribution}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}
}
\value{
Vector
}
\description{
Weibull-with-p1 quantile function
}
