% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02a_reltest.R
\name{reltest}
\alias{reltest}
\title{Evaluation of Reliability for Models in the \code{fitdistcp} Package}
\usage{
reltest(
  model = "exp",
  ntrials = 1000,
  nrepeats = 3,
  nx = 20,
  params = c(1),
  alpha = seq(0.005, 0.995, 0.005),
  plotflag = TRUE,
  verbose = TRUE,
  dmgs = TRUE,
  debug = FALSE,
  aderivs = TRUE,
  unbiasedv = FALSE,
  pwm = FALSE
)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{exp}",
"\code{pareto_k1}",
"\code{halfnorm}",
"\code{unif}",
"\code{norm}",
"\code{norm_dmgs}",
"\code{gnorm_k3}",
"\code{lnorm}",
"\code{lnorm_dmgs}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{exp_p1}",
"\code{pareto_p1k3}",
"\code{norm_p1}",
"\code{lnorm_p1}",
"\code{logis_p1}",
"\code{lst_p1k4}",
"\code{cauchy_p1}",
"\code{gumbel_p1}",
"\code{frechet_p2k1}",
"\code{weibull_p2}",
"\code{gev_p1k4}",
"\code{norm_p12}",
"\code{lst_p12k5}",
"\code{gamma}",
"\code{invgamma}",
"\code{invgauss}",
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".
"\code{gev_p12}".
"\code{gev_p123}".}

\item{ntrials}{the number of trials to run. 5000 typically gives good results.}

\item{nrepeats}{the number of entire repeats of the test to run, to check for convergence. 3 is a good choice.}

\item{nx}{the length of the training data to use.}

\item{params}{values for the parameters for the specified distribution}

\item{alpha}{the exceedance probability values at which to test}

\item{plotflag}{logical to turn the plotting on and off}

\item{verbose}{logical to turn loop counting on and off}

\item{dmgs}{logical to turn DMGS calculations on and off (to optimize speed for maxlik only calculations)}

\item{debug}{flag for turning debug messages on}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}

\item{unbiasedv}{a logical for whether to use the unbiased variance instead of maxlik (for the normal)}

\item{pwm}{a logical for whether to use PWM instead of maxlik (for the GEV)}
}
\value{
A plot showing 9 different diagnostic checks, and a list containing
various outputs, including the
probabilities shown in the plot.
}
\description{
Uses simulations to evaluate the reliability of
the predictive quantiles produced by the \code{q****_cp} routines in the \code{fitdistcp} package.
}
\details{
The maximum likelihood quantiles (plotted in blue) do not give good reliability.
They typically underestimate the tails (see panel (f)).

For
"\code{exp}",
"\code{pareto_k1}",
"\code{unif}",
"\code{norm}",
"\code{lnorm}",
"\code{norm_p1}" and
"\code{lnorm_p1}",
the calibrating prior quantiles are calculated using the right Haar prior
and an exact solution for the Bayesian prediction integral.
They will converge towards exact reliability with a large enough number of trials,
for any sample size.

For
"\code{halfnorm}",
"\code{norm_dmgs}",
"\code{lnorm_dmgs}",
"\code{gnorm_k3}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{exp_p1}",
"\code{pareto_p1k3}",
"\code{gumbel_p1}",
"\code{logis_p1}" and
"\code{lst_p1k4}"
"\code{cauchy_p1}",
"\code{gumbel_p1}",
"\code{frechet_p2k1}",
"\code{weibull_p2}",
"\code{gev_p1k4}",
"\code{norm_p12}",
"\code{lst_p12k5}"
the calibrating prior quantiles are calculated using the right Haar prior,
with the DMGS asymptotic solution for the Bayesian prediction integral.
They will converge towards good reliability with a large enough number of trials,
with the only deviation from exact reliability being due to the neglect of
higher order terms in the asymptotic expansion.
They will converge towards exact reliability with a large enough number of trials
and a large enough sample size.

For
"\code{gamma}",
"\code{invgamma}",
"\code{invgauss}",
"\code{gev}",
"\code{gpd_k1}" and
"\code{gev_p1}",
"\code{gev_p12}",
"\code{gev_p123}",
the calibrating prior quantiles are calculated using the recommend priors,
with the DMGS asymptotic solution for the Bayesian prediction integral.
The chosen priors give reasonably good reliability with a
large enough number of trials,
and for large sample sizes, but may give poor reliability for small
sample sizes (e.g., n<20).
}
\examples{
set.seed(1)
# example 1
# -runs the default settings, which test reliability for the exponential distribution
reltest()

}
\references{
If you use this package, we would be grateful if you would cite the following reference,
which gives the various calibrating priors, and tests them for reliability:

\itemize{
\item Jewson S., Sweeting T. and Jewson L. (2024): Reducing Reliability Bias in
Assessments of Extreme Weather Risk using Calibrating Priors;
ASCMO (Advances in Statistical Climatology, Meteorology and Oceanography)
}

The proof that using the right Haar prior gives exact probability
matching for homogeneous models is given in:
\itemize{
\item Severini, T., Mukerjee, R. and Ghosh, M. (2002); Biometrika
}

The DMGS asymptotic expansions come from:
\itemize{
\item Datta, G., Mukerjee, R., Ghosh, M. and Sweeting, T. (2000); Annals of Statistics
}
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
