% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_reltest2.R
\name{reltest2}
\alias{reltest2}
\title{Evaluation of Reliability for Certain Additional Models in the \code{fitdistcp} Package}
\usage{
reltest2(
  model = "gev",
  ntrials = 100,
  nrepeats = 3,
  nx = 50,
  params = c(0, 1, 0),
  alpha = seq(0.005, 0.995, 0.005),
  plotflag = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".}

\item{ntrials}{the number of trials to run. 5000 typically gives good results.}

\item{nrepeats}{the number of entire repeats of the test to run, to check for convergence. 3 is a good choice.}

\item{nx}{the length of the training data.}

\item{params}{values for the parameters for the specified distribution}

\item{alpha}{the alpha values at which to test}

\item{plotflag}{logical to turn the plotting on and off}

\item{verbose}{logical to turn loop counting on and off}
}
\value{
A plot showing 9 different diagnostic checks, and a list containing
various outputs, including the
probabilities shown in the plot.
}
\description{
This routine is mainly for reproducing certain results in Jewson et al. (2024),
and not of general interest.

It uses simulations to evaluate the reliability of
the predictive quantiles produced by the
\code{qgev_cp}, \code{ggpd_cp} and \code{qgev_p1_cp}
routines in the \code{fitdistcp} package.
For each model, results for 5 models are calculated.
This is to illustrate that the calibrating prior predictions dominate
the \code{ml}, \code{flat}, \code{crhp_ml} and \code{jp} predictions,
in terms of reliability.
}
\details{
The maximum likelihood quantiles (plotted in blue) do not give good reliability. They typically underestimate the tails (see panel (f)).

The \code{cp} predictive quantiles generally give reasonably good reliability, especially for sample sizes of ~100.
The other predictions generally give poor reliability.
}
\examples{
set.seed(1)
# example 1
# -runs the default settings, which test reliability for the GEV distribution
reltest2()
}
\references{
If you use this package, we would be grateful if you would cite the following reference,
which gives the various calibrating priors, and tests them for reliability:

\itemize{
\item Jewson S., Sweeting T. and Jewson L. (2024): Reducing Reliability Bias in
Assessments of Extreme Weather Risk using Calibrating Priors;
ASCMO (Advances in Statistical Climatology, Meteorology and Oceanography)
}

The proof that using the right Haar prior gives exact probability
matching for homogeneous models is given in:
\itemize{
\item Severini, T., Mukerjee, R. and Ghosh, M. (2002); Biometrika
}

The DMGS asymptotic expansions come from:
\itemize{
\item Datta, G., Mukerjee, R., Ghosh, M. and Sweeting, T. (2000); Annals of Statistics
}
}
\author{
Stephen Jewson \email{stephen.jewson@gmail.com}
}
