% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_reltest2_libs.R
\name{reltest2_cases}
\alias{reltest2_cases}
\title{Cases}
\usage{
reltest2_cases(model = "gev", nx = 50, params)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"gev",
"gpd_k1",
"gev_pred1".}

\item{nx}{length of training data}

\item{params}{model parameters}
}
\value{
Two integers
}
\description{
Cases
}
