% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_reltest2_libs.R
\name{reltest2_makeep}
\alias{reltest2_makeep}
\title{Cases}
\usage{
reltest2_makeep(model, pred1, tt0, params)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"gev",
"gpd_k1",
"gev_pred1".}

\item{pred1}{quantile predictions}

\item{tt0}{value of predictor vector}

\item{params}{model parameters}
}
\value{
Vector
}
\description{
Cases
}
