% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003c_reltest2_plot_libs.R
\name{reltest2_plot}
\alias{reltest2_plot}
\title{Plotting routine for reltest2}
\usage{
reltest2_plot(
  model,
  ntrials,
  nrepeats,
  nx,
  params,
  nmethods,
  alpha,
  freqexceeded,
  case
)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"gev",
"gpd",
"gev_p1".}

\item{ntrials}{the number of trials o run. 5000 typically gives good results.}

\item{nrepeats}{the number of entire repeats of the test to run, to check for convergence}

\item{nx}{the length of the training data.}

\item{params}{values for the parameters for the specified distribution}

\item{nmethods}{the number of methods being tested}

\item{alpha}{the values of alpha being tested}

\item{freqexceeded}{the exceedance counts}

\item{case}{there are 3 cases (must be set to case=1 except for my testing)}
}
\value{
Plots the results of reliability testing
}
\description{
Plots 9 diagnostics related to predictive probability matching.
}
