% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_reltest2_libs.R
\name{reltest2_predict}
\alias{reltest2_predict}
\title{Make prediction from one model}
\usage{
reltest2_predict(model = "gev", xx, tt, n0, pp, params, case, nmethods)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{exp}",
"\code{pareto_k1}",
"\code{halfnorm}",
"\code{norm}",
"\code{lnorm}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{norm_p1}",
"\code{lnorm_p1}",
"\code{logis_p1}",
"\code{lst_k3p1}",
"\code{gumbel_p1}",
"\code{norm_p12}",
"\code{gev}",
"\code{gpd}",
"\code{gev_p1}".}

\item{xx}{training data}

\item{tt}{predictor vector}

\item{n0}{index for predictor vector}

\item{pp}{probabilities to predict}

\item{params}{model parameters}

\item{case}{the case number: different models have different lists of methods}

\item{nmethods}{the number of methods: different models have different numbers of methods}
}
\description{
Make prediction from one model
}
