% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002b_reltest_libs.R
\name{reltest_makeep}
\alias{reltest_makeep}
\title{Calculate EP from one model}
\usage{
reltest_makeep(model, pred1, tt0, tt10, tt20, tt30, params)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{exp}",
"\code{pareto_k2}",
"\code{halfnorm}",
"\code{unif}",
"\code{norm}",
"\code{norm_dmgs}",
"\code{gnorm_k3}",
"\code{lnorm}",
"\code{lnorm_dmgs}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{exp_p1}",
"\code{pareto_p1k2}",
"\code{norm_p1}",
"\code{lnorm_p1}",
"\code{logis_p1}",
"\code{lst_p1k3}",
"\code{cauchy_p1}",
"\code{gumbel_p1}",
"\code{frechet_p2k1}",
"\code{weibull_p2}",
"\code{gev_p1k3}",
"\code{norm_p12}",
"\code{lst_p12k3}",
"\code{gamma}",
"\code{invgamma}",
"\code{invgauss}",
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".
"\code{gev_p12}".
"\code{gev_p123}".}

\item{pred1}{quantile predictions}

\item{tt0}{value of the predictor}

\item{tt10}{value of predictor 1}

\item{tt20}{value of predictor 2}

\item{tt30}{value of predictor 3}

\item{params}{the model parameters}
}
\value{
Vector
}
\description{
Calculate EP from one model
}
