% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_reltest_libs.R
\name{reltest_makemaxep}
\alias{reltest_makemaxep}
\title{Calculate MaxEP from one model}
\usage{
reltest_makemaxep(model, ml_max, tt0, tt10, tt20, tt30, params)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".
"\code{gev_p12}".
"\code{gev_p123}".}

\item{ml_max}{predicted max value}

\item{tt0}{value of the predictor}

\item{tt10}{value of predictor 1}

\item{tt20}{value of predictor 2}

\item{tt30}{value of predictor 3}

\item{params}{the model parameters}
}
\value{
Vector
}
\description{
Calculate MaxEP from one model
}
