% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_reltest_libs.R
\name{reltest_predict}
\alias{reltest_predict}
\title{Make prediction from one model}
\usage{
reltest_predict(
  model,
  xx,
  tt,
  tt1,
  tt2,
  tt3,
  n0,
  n10,
  n20,
  n30,
  pp,
  params,
  dmgs = TRUE,
  debug = FALSE,
  aderivs = TRUE,
  unbiasedv = FALSE,
  pwm = FALSE,
  minxi = -10,
  maxxi = 10
)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{exp}",
"\code{pareto_k2}",
"\code{halfnorm}",
"\code{unif}",
"\code{norm}",
"\code{norm_dmgs}",
"\code{gnorm_k3}",
"\code{lnorm}",
"\code{lnorm_dmgs}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{exp_p1}",
"\code{pareto_p1k2}",
"\code{norm_p1}",
"\code{lnorm_p1}",
"\code{logis_p1}",
"\code{lst_p1k3}",
"\code{cauchy_p1}",
"\code{gumbel_p1}",
"\code{frechet_p2k1}",
"\code{weibull_p2}",
"\code{exp_p1k4}",
"\code{norm_p12}",
"\code{lst_p12k3}",
"\code{gamma}",
"\code{invgamma}",
"\code{invgauss}",
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".
"\code{gev_p12}".
"\code{gev_p123}".}

\item{xx}{training data}

\item{tt}{predictor vector}

\item{tt1}{predictor vector 1}

\item{tt2}{predictor vector 2}

\item{tt3}{predictor vector 3}

\item{n0}{index for predictor vector}

\item{n10}{index for predictor vector 1}

\item{n20}{index for predictor vector 2}

\item{n30}{index for predictor vector 2}

\item{pp}{probabilites at which to make quantile predictions}

\item{params}{model parameters}

\item{dmgs}{flag for whether to run dmgs calculations or not}

\item{debug}{flag for turning debug messages on}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}

\item{unbiasedv}{a logical for whether to use the unbiased variance instead of maxlik (for the normal)}

\item{pwm}{a logical for whether to use PWM instead of maxlik (for the GEV)}

\item{minxi}{minimum value for EVT shape parameter}

\item{maxxi}{maximum value for EVT shape parameter}
}
\value{
Two vectors
}
\description{
Make prediction from one model
}
