% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_reltest_libs.R
\name{reltest_simulate}
\alias{reltest_simulate}
\title{Random training data from one model}
\usage{
reltest_simulate(model = "exp", nx = 20, tt, tt1, tt2, tt3, params)
}
\arguments{
\item{model}{which distribution to test. Possibles values are
"\code{exp}",
"\code{pareto_k2}",
"\code{halfnorm}",
"\code{unif}",
"\code{norm}",
"\code{norm_dmgs}",
"\code{gnorm_k3}",
"\code{lnorm}",
"\code{lnorm_dmgs}",
"\code{logis}",
"\code{lst_k3}",
"\code{cauchy}",
"\code{gumbel}",
"\code{frechet_k1}",
"\code{weibull}",
"\code{gev_k3}",
"\code{exp_p1}",
"\code{pareto_p1k2}",
"\code{norm_p1}",
"\code{lnorm_p1}",
"\code{logis_p1}",
"\code{lst_p1k3}",
"\code{cauchy_p1}",
"\code{gumbel_p1}",
"\code{frechet_p2k1}",
"\code{weibull_p2}",
"\code{gev_p1k3}",
"\code{norm_p12}",
"\code{lst_p12k3}",
"\code{gamma}",
"\code{invgamma}",
"\code{invgauss}",
"\code{gev}",
"\code{gpd_k1}",
"\code{gev_p1}".
"\code{gev_p12}".
"\code{gev_p123}".}

\item{nx}{the length of the training data to use.}

\item{tt}{predictor vector}

\item{tt1}{predictor vector 1}

\item{tt2}{predictor vector 2}

\item{tt3}{predictor vector 2}

\item{params}{values for the parameters for the specified distribution}
}
\description{
Random training data from one model
}
