% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{rgev_p123_minmax}
\alias{rgev_p123_minmax}
\title{rgev for gev_p123 but with maxlik xi within bounds}
\usage{
rgev_p123_minmax(
  nx,
  mu,
  sigma,
  xi,
  t1,
  t2,
  t3,
  minxi = -0.45,
  maxxi = 0.45,
  centering = TRUE
)
}
\arguments{
\item{nx}{length of training data}

\item{mu}{the location parameter of the distribution}

\item{sigma}{the sigma parameter of the distribution}

\item{xi}{the shape parameter of the distribution}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{centering}{indicates whether the routine should center the data or not}
}
\value{
Vector
}
\description{
rgev for gev_p123 but with maxlik xi within bounds
}
