% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02c_reltest_plot_libs.R
\name{testppm_plot}
\alias{testppm_plot}
\title{Plotting routine for testppm}
\usage{
testppm_plot(
  model,
  ntrials,
  nrepeats,
  nx,
  params,
  nmethods,
  alpha,
  freqexceeded
)
}
\arguments{
\item{model}{which distribution to test. Possibles values are}

\item{ntrials}{the number of trials to run. 5000 typically gives good results.}

\item{nrepeats}{the number of entire repeats of the test to run, to check for convergence}

\item{nx}{the length of the training data.}

\item{params}{values for the parameters for the specified distribution}

\item{nmethods}{the number of methods being tested}

\item{alpha}{the values of alpha being tested}

\item{freqexceeded}{the exceedance counts}
}
\value{
Plots the results of reliability testing
}
\description{
Plots 9 diagnostics related to predictive probability matching.
}
