% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52c_weibull_derivs.R
\name{weibull_f1fa}
\alias{weibull_f1fa}
\title{The first derivative of the density}
\usage{
weibull_f1fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The first derivative of the density
}
