% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52c_weibull_derivs.R
\name{weibull_fdd}
\alias{weibull_fdd}
\title{Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
weibull_fdd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Second derivative of the density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
