% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52b_weibull_libs.R
\name{weibull_ldd}
\alias{weibull_ldd}
\title{Second derivative matrix of the normalized log-likelihood}
\usage{
weibull_ldd(x, v1, fd1, v2, fd2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}
}
\description{
Second derivative matrix of the normalized log-likelihood
}
