% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052b_weibull_libs.R
\name{weibull_lmnp}
\alias{weibull_lmnp}
\title{One component of the third derivative of the normalized log-likelihood}
\usage{
weibull_lmnp(x, v1, fd1, v2, fd2, mm, nn, rr)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}

\item{rr}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the third derivative of the normalized log-likelihood
}
