% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052b_weibull_libs.R
\name{weibull_logf}
\alias{weibull_logf}
\title{Logf for RUST}
\usage{
weibull_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
Logf for RUST
}
