% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52c_weibull_derivs.R
\name{weibull_logfddd}
\alias{weibull_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
weibull_logfddd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
