% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52b_weibull_libs.R
\name{weibull_loglik}
\alias{weibull_loglik}
\title{log-likelihood function}
\usage{
weibull_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\description{
log-likelihood function
}
