% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073b_weibull_p2_libs.R
\name{weibull_p2_lddd}
\alias{weibull_p2_lddd}
\title{Third derivative tensor of the normalized log-likelihood}
\usage{
weibull_p2_lddd(x, t, v1, fd1, v2, d2, v3, d3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}
}
\value{
Cubic scalar array
}
\description{
Third derivative tensor of the normalized log-likelihood
}
