% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073b_weibull_p2_libs.R
\name{weibull_p2_lmn}
\alias{weibull_p2_lmn}
\title{One component of the second derivative of the normalized log-likelihood}
\usage{
weibull_p2_lmn(x, t, v1, fd1, v2, d2, v3, d3, mm, nn)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{v2}{second parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{v3}{third parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{mm}{an index for which derivative to calculate}

\item{nn}{an index for which derivative to calculate}
}
\value{
Scalar value
}
\description{
One component of the second derivative of the normalized log-likelihood
}
