% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73b_weibull_p2_libs.R
\name{weibull_p2_loglik}
\alias{weibull_p2_loglik}
\title{observed log-likelihood function}
\usage{
weibull_p2_loglik(vv, x, t)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\description{
observed log-likelihood function
}
