% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73b_weibull_p2_libs.R
\name{weibull_p2_means}
\alias{weibull_p2_means}
\title{weibull distribution: RHP mean}
\usage{
weibull_p2_means(means, t0, ml_params, lddi, lddd, lambdad_rhp, nx, dim)
}
\arguments{
\item{means}{logical that indicates whether to return analytical estimates for the distribution means (longer runtime)}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{ml_params}{parameters}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad_rhp}{derivative of the log RHP prior}

\item{nx}{length of training data}

\item{dim}{number of parameters}
}
\value{
Two scalars
}
\description{
weibull distribution: RHP mean
}
