% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73c_weibull_p2_derivs.R
\name{weibull_p2_mu1fa}
\alias{weibull_p2_mu1fa}
\title{Minus the first derivative of the cdf, at alpha}
\usage{
weibull_p2_mu1fa(alpha, t, v1, v2, v3)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
Minus the first derivative of the cdf, at alpha
}
