% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73c_weibull_p2_derivs.R
\name{weibull_p2_p1fa}
\alias{weibull_p2_p1fa}
\title{The first derivative of the cdf}
\usage{
weibull_p2_p1fa(x, t, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\description{
The first derivative of the cdf
}
