% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073c_weibull_p2_derivs.R
\name{weibull_p2_p2fa}
\alias{weibull_p2_p2fa}
\title{The second derivative of the cdf}
\usage{
weibull_p2_p2fa(x, t0, v1, v2, v3)
}
\arguments{
\item{x}{a vector of training data values}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{v1}{first parameter}

\item{v2}{second parameter}

\item{v3}{third parameter}
}
\value{
Matrix
}
\description{
The second derivative of the cdf
}
