% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/73b_weibull_p2_libs.R
\name{weibull_p2_predictordata}
\alias{weibull_p2_predictordata}
\title{Predicted Parameter and Generalized Residuals}
\usage{
weibull_p2_predictordata(predictordata, x, t, t0, params)
}
\arguments{
\item{predictordata}{a logical that indicates whether to calculate and return predictordata}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{t0}{a single value of the predictor (specify either \code{t0} or \code{n0} but not both)}

\item{params}{model parameters for calculating logf}
}
\description{
Predicted Parameter and Generalized Residuals
}
