% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52c_weibull_derivs.R
\name{weibull_p2fa}
\alias{weibull_p2fa}
\title{The second derivative of the cdf}
\usage{
weibull_p2fa(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\description{
The second derivative of the cdf
}
