% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052c_weibull_derivs.R
\name{weibull_pd}
\alias{weibull_pd}
\title{First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
weibull_pd(x, v1, v2)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}

\item{v2}{second parameter}
}
\value{
Vector
}
\description{
First derivative of the cdf
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
