% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42b_cauchy_libs.R
\name{cauchy_logf}
\alias{cauchy_logf}
\title{Logf for RUST}
\usage{
cauchy_logf(params, x)
}
\arguments{
\item{params}{model parameters for calculating logf}

\item{x}{a vector of training data values}
}
\description{
Logf for RUST
}
