% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051b_frechet_k1_libs.R
\name{dfrechetsub}
\alias{dfrechetsub}
\title{Densities from MLE and RHP}
\usage{
dfrechetsub(x, y, kloc)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{kloc}{the known location parameter}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
