% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{dgev_p123sub}
\alias{dgev_p123sub}
\title{Densities for 5 predictions}
\usage{
dgev_p123sub(x, t1, t2, t3, y, t01, t02, t03, ics, extramodels, debug)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{t03}{a single value of the predictor (specify either \code{t03} or \code{n03} but not both)}

\item{ics}{initial conditions for the maximum likelihood search}

\item{extramodels}{logical that indicates whether to add three additional prediction models}

\item{debug}{debug flag}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities for 5 predictions
}
