% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031b_norm_dmgs_libs.R
\name{dnorm_dmgssub}
\alias{dnorm_dmgssub}
\title{Densities from MLE and RHP}
\usage{
dnorm_dmgssub(x, y)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities from MLE and RHP
}
