% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{dnorm_p12dmgs}
\alias{dnorm_p12dmgs}
\title{Densities for 5 predictions}
\usage{
dnorm_p12dmgs(
  x,
  t1,
  t2,
  y,
  t10,
  t20,
  ics,
  d1 = 0.01,
  d2 = 0.01,
  d3 = 0.01,
  d4 = 0.01,
  aderivs = TRUE
)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{t10}{a single value of the predictor for the mean (specify either \code{t10} or \code{n10} but not both)}

\item{t20}{a single value of the predictor for the sd (specify either \code{t20} or \code{n20} but not both)}

\item{ics}{initial conditions for the maximum likelihood search}

\item{d1}{the delta used in the numerical derivatives with respect to the parameter}

\item{d2}{the delta used in the numerical derivatives with respect to the parameter}

\item{d3}{the delta used in the numerical derivatives with respect to the parameter}

\item{d4}{the delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{logical for whether to use analytic derivatives (instead of numerical)}
}
\value{
A vector of parameter estimates, two pdf vectors, two cdf vectors
}
\description{
Densities for 5 predictions
}
