% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11b_pareto_k2_libs.R
\name{dpareto_k2_sub}
\alias{dpareto_k2_sub}
\title{Densities from MLE and RHP}
\usage{
dpareto_k2_sub(x, y, kscale, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{kscale}{the known scale parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Densities from MLE and RHP
}
