% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/52b_weibull_libs.R
\name{dweibullsub}
\alias{dweibullsub}
\title{Densities from MLE and RHP}
\usage{
dweibullsub(x, y, fd1 = 0.01, fd2 = 0.01, aderivs = TRUE)
}
\arguments{
\item{x}{a vector of training data values}

\item{y}{a vector of values at which to calculate the density and distribution functions}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{fd2}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Densities from MLE and RHP
}
