% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010c_exp_derivs.R, R/10c_exp_derivs.R
\name{exp_f2fa}
\alias{exp_f2fa}
\title{The second derivative of the density}
\usage{
exp_f2fa(x, v1)

exp_f2fa(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\value{
Matrix

Matrix
}
\description{
The second derivative of the density

The second derivative of the density
}
