% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10c_exp_derivs.R
\name{exp_logfddd}
\alias{exp_logfddd}
\title{Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol}
\usage{
exp_logfddd(x, v1)
}
\arguments{
\item{x}{a vector of training data values}

\item{v1}{first parameter}
}
\description{
Third derivative of the log density
Created by Stephen Jewson
using Deriv() by Andrew Clausen and Serguei Sokol
}
