% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110b_gev_libs.R, R/160b_gev_extramodels_libs.R
\name{gev_loglik}
\alias{gev_loglik}
\title{log-likelihood function}
\usage{
gev_loglik(vv, x)

gev_loglik(vv, x)
}
\arguments{
\item{vv}{parameters}

\item{x}{a vector of training data values}
}
\value{
Scalar value.
}
\description{
log-likelihood function

log-likelihood function
}
