% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/152b_gev_p123_libs.R
\name{gev_p123_setics}
\alias{gev_p123_setics}
\title{Set initial conditions}
\usage{
gev_p123_setics(x, t1, t2, t3, ics)
}
\arguments{
\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{t3}{a vector of predictors for the shape}

\item{ics}{initial conditions for the maximum likelihood search}
}
\description{
Set initial conditions
}
