% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/154b_gev_p12_libs.R
\name{gev_p12_waic}
\alias{gev_p12_waic}
\title{Waic}
\usage{
gev_p12_waic(
  waicscores,
  x,
  t1,
  t2,
  v1hat,
  v2hat,
  v3hat,
  v4hat,
  v5hat,
  lddi,
  lddd,
  lambdad
)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}

\item{v3hat}{third parameter}

\item{v4hat}{fourth parameter}

\item{v5hat}{fifth parameter}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad}{derivative of the log prior}
}
\value{
Two numeric values.
}
\description{
Waic
}
