% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20b_halfnorm_libs.R
\name{halfnorm_logscores}
\alias{halfnorm_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
halfnorm_logscores(logscores, x, fd1 = 0.01, aderivs = TRUE)
}
\arguments{
\item{logscores}{a logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{fd1}{the fractional delta used in the numerical derivatives with respect to the parameter}

\item{aderivs}{a logical for whether to use analytic derivatives (instead of numerical)}
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
