% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061b_lnorm_p1_libs.R
\name{lnorm_p1_waic}
\alias{lnorm_p1_waic}
\title{Waic}
\usage{
lnorm_p1_waic(waicscores, x, t, v1hat, v2hat, v3hat)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}

\item{v3hat}{third parameter}
}
\value{
Two numeric values.
}
\description{
Waic
}
