% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{make_se}
\alias{make_se}
\title{Make Standad Errors from lddi}
\usage{
make_se(nx, lddi)
}
\arguments{
\item{nx}{length of training data}

\item{lddi}{the inverse log-likelihood matrix}
}
\description{
Make Standad Errors from lddi
}
