% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{norm_p12_logscores}
\alias{norm_p12_logscores}
\title{Log scores for 5 predictions calculated using leave-one-out}
\usage{
norm_p12_logscores(logscores, x, t1, t2, ics)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t1}{a vector of predictors for the mean}

\item{t2}{a vector of predictors for the sd}

\item{ics}{initial conditions for the maximum likelihood search}
}
\value{
Two scalars
}
\description{
Log scores for 5 predictions calculated using leave-one-out
}
