% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011c_pareto_k2_derivs.R,
%   R/11c_pareto_k2_derivs.R
\name{pareto_k2_mu2fa}
\alias{pareto_k2_mu2fa}
\title{Minus the second derivative of the cdf, at alpha}
\usage{
pareto_k2_mu2fa(alpha, v1, kscale)

pareto_k2_mu2fa(alpha, v1, kscale)
}
\arguments{
\item{alpha}{a vector of values of alpha (one minus probability)}

\item{v1}{first parameter}

\item{kscale}{the known scale parameter}
}
\value{
Matrix

Matrix
}
\description{
Minus the second derivative of the cdf, at alpha

Minus the second derivative of the cdf, at alpha
}
