% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080b_norm_p12_libs.R
\name{qnorm_p12}
\alias{qnorm_p12}
\title{Normal-with-p12: Quantile function}
\usage{
qnorm_p12(p, t01, t02, ymn, slope, sigma1, sigma2)
}
\arguments{
\item{p}{a vector of probabilities at which to generate predictive quantiles}

\item{t01}{a single value of the predictor (specify either \code{t01} or \code{n01} but not both)}

\item{t02}{a single value of the predictor (specify either \code{t02} or \code{n02} but not both)}

\item{ymn}{the location parameter of the function of the predictor}

\item{slope}{the slope of the function of the predictor}

\item{sigma1}{first coefficient for the sigma parameter of the distribution}

\item{sigma2}{second coefficient for the sigma parameter of the distribution}
}
\value{
Vector
}
\description{
Normal-with-p12: Quantile function
}
