% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/150b_gev_p1_libs.R
\name{rgev_p1_minmax}
\alias{rgev_p1_minmax}
\title{rgev for gev_p1 but with maxlik xi within bounds}
\usage{
rgev_p1_minmax(
  nx,
  mu,
  sigma,
  xi,
  tt,
  minxi = -0.45,
  maxxi = 0.45,
  centering = TRUE
)
}
\arguments{
\item{nx}{length of training data}

\item{mu}{the location parameter of the distribution}

\item{sigma}{the sigma parameter of the distribution}

\item{xi}{the shape parameter of the distribution}

\item{tt}{a vector of predictors}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}

\item{centering}{indicates whether the routine should center the data or not}
}
\value{
Vector
}
\description{
rgev for gev_p1 but with maxlik xi within bounds
}
