% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120b_gpd_k1_libs.R
\name{rgpd_k1_minmax}
\alias{rgpd_k1_minmax}
\title{rgpd for gpd_k1 but with maxlik xi within bounds}
\usage{
rgpd_k1_minmax(nx, kloc, sigma, xi, minxi = -0.45, maxxi = 0.45)
}
\arguments{
\item{nx}{length of training data}

\item{kloc}{the known location parameter}

\item{sigma}{the sigma parameter of the distribution}

\item{xi}{the shape parameter of the distribution}

\item{minxi}{minimum value of shape parameter xi}

\item{maxxi}{maximum value of shape parameter xi}
}
\value{
Vector
}
\description{
rgpd for gpd_k1 but with maxlik xi within bounds
}
