% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073b_weibull_p2_libs.R
\name{weibull_p2_logscores}
\alias{weibull_p2_logscores}
\title{Log scores for MLE and RHP predictions calculated using leave-one-out}
\usage{
weibull_p2_logscores(logscores, x, t)
}
\arguments{
\item{logscores}{logical that indicates whether to return leave-one-out estimates estimates of the log-score (much longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}
}
\value{
Two scalars
}
\description{
Log scores for MLE and RHP predictions calculated using leave-one-out
}
