% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073b_weibull_p2_libs.R
\name{weibull_p2_waic}
\alias{weibull_p2_waic}
\title{Waic}
\usage{
weibull_p2_waic(waicscores, x, t, v1hat, v2hat, v3hat, lddi, lddd, lambdad)
}
\arguments{
\item{waicscores}{logical that indicates whether to return estimates for the waic1 and waic2 scores (longer runtime)}

\item{x}{a vector of training data values}

\item{t}{a vector or matrix of predictors}

\item{v1hat}{first parameter}

\item{v2hat}{second parameter}

\item{v3hat}{third parameter}

\item{lddi}{inverse observed information matrix}

\item{lddd}{third derivative of log-likelihood}

\item{lambdad}{derivative of the log prior}
}
\value{
Two numeric values.
}
\description{
Waic
}
